Only create the abi that we are going to use
Summary:
In Java we can create three different types of ABI - class, source and source-only. When we're doing an Android build, we only ever need one of these, so doing analysis for any others is a waste of time as it will never be used.

We still want all of them available as e.g. sub_targets on a build of the library itself, so we keep them all around in that case.

Reviewed By: jkeljo

Differential Revision: D40304565

fbshipit-source-id: 2add4b9716a6429dd1b1fd36cd73eaaeeb8c1c6d