Don't create AST subtarget when building an Android binary
Summary:
When we're building an Android binary we never need to use this subtarget, and the retained memory from analysis is about 0.5G for all of these subtargets. Therefore just don't create them when we're building an Android binary.

In the previous diff we added the configuration and attributes to specify that we're building an Android binary, this just wires it up.

Reviewed By: jkeljo

Differential Revision: D40336765

fbshipit-source-id: 87f9948a5e902a6f911a8de3e75b27aecda5cd33