Allow tagging only input artifacts
Summary:
When we're tagging things for dep-files, we only want one output - the dep file itself. If we have a command line that includes inputs and outputs, we might therefore only want to tag the inputs.

(May not be necessary after T134944772, but seems a reasonable thing to have in general).

Reviewed By: krallin

Differential Revision: D40381989

fbshipit-source-id: c88dc8884fd7eb687404a647df088586504e176a