Fix re_uploads
Summary:
This test was failign due to buck2 usage of ```common/rust/shed/hostname``` which used FbWhoAmI and within it has a hardcoded path that doesn't work on windows corp machines.

To fix this we simply switch crates to [this one](https://crates.io/crates/hostname) and change the sh script for a python one.

Reviewed By: krallin

Differential Revision: D40175680

fbshipit-source-id: 0776711f22d2424bfe7019fb37209fce9d5b0663