Do not remove buckd.lifecycle in clean command
Summary:
mariof reported that
* home dir is on NFS
* `remove_dir_all` fails to remove daemon dir every time `buck2 clean` is invoked

`remove_dir_all` does not explain which file it failed to remove exactly, but it is good evidence that it is `buckd.lifecycle`.

So do not remove `buckd.lifecycle` in `buck2 clean`.

Seems to be the right thing to do anyway: not remove the lock while holding the lock.

This way, `buck2 clean` won't clean all of what it was cleaned before, but that should be fine.

Reviewed By: krallin

Differential Revision: D40423021

fbshipit-source-id: 57f10087dfab3d4d68a6d1656e8c830589acca3e