buck2: dice: look for force dirtied across all non-verified versions
Summary:
like it says in the title. Considered the following sequence of events in
Buck2.

- Build 1 reads file A.
- Build 2 changes config, but doesn't read A.
  - This dirties the cell resolver and thus dirties A (but not forced)
- User changes A
- Build 3 reads files A
  - This build first force-dirties A then runs

What will currently happen is we ignore the force-dirty since we see a dirtied
version earlier. As a result we get a broken build:

https://fb.workplace.com/groups/buck2users/permalink/3264459227143769/

This diff fixes it by looking for a force-dirtied version across all versions.

Reviewed By: bobyangyf

Differential Revision: D40428261

fbshipit-source-id: 915e0a52459d0cda438aa330b2cf683cd229dc45