Add used_classes_to_dep_file tool to Java toolchain
Summary:
The format for dep_files in buck2 is for it to just be a list of files that were used in compilation. Right now javacd dumps a JSON mapping of jars to the files that were used inside the jar.

We just want to the jar paths, so add this wrapper that will run javacd and then convert the JSON to a list of jar paths.

We also need to write some paths that we know are used (due to T134944772), so allow that too.

Reviewed By: milend

Differential Revision: D40382763

fbshipit-source-id: 3e8a1eaddb02cde2ef4f9fae96b9c163205f74ac