Only do cxx_library work that we actually need for Android builds
Summary:
By default, the `cxx_library` rule will produce all possible providers and sub targets. But on Android we don't actually need a bunch of those, so let's tell the rule not to produce those, since it's just wasted work.

Would be interested if anyone else can see (perhaps from the profile in the test plan) if we're doing any other work that we don't need to.

Reviewed By: smeenai

Differential Revision: D40102858

fbshipit-source-id: f4a3f899b04c85387e14a33d4c0b44dbfe86ac96