buck2: add `force_full_hybrid_if_capable` on actions
Summary:
This is spiritually a bit like `prefer_local`, but with regard to limited vs. full
hybrid. We'd like to have this because some actions are running locally today,
could run remotely, but would potentially get slower if they did that (because
large actions has more overhead on RE), unless we have a lot of actions, in
which case we win by running on RE.

This is the kind of scenario that hybrid allows us to optimize (though it is
currently a pretty coarse instrument), so let us have those specific commands
opt into this.

Note that long term it's not clear that we ever want to have full hybrid on CI
since we'd lose sandboxing, so for now this is really for things that otherwise
would be `prefer_local = True`.

Reviewed By: johnny90

Differential Revision: D40434517

fbshipit-source-id: c8f30f3e6b3feec3b1d0c3aad7923fd6e740579d