Log which trace ID is being blocked
Summary:
Log which trace ID is being blocked, and which is currently executing. Also change the trace to a `warn` so that it's visible by default to the users.

I unwrap the Option without checking, since it should always be populated by the time the parallel command blocking logic is encountered.

Reviewed By: ndmitchell

Differential Revision: D40432050

fbshipit-source-id: 64df5ff90074243204af45b10d58fdbfc776e862