Log buck2 revision/version information in buck2_action_errors Scuba table
Summary:
Tracking the buck2 version across builds allows us to determine which versions introduce regressions. Currently, the only other Scuba table that tracks this is `buck2_builds`.

This adds the buck2 revision hash to `buck2_action_errors`.

The field is labeled `optional` in the `ActionExecutionEnd` protobuf message because we are only logging it in the case of failed action, and not a successful one.

Reviewed By: krallin

Differential Revision: D39950206

fbshipit-source-id: 6b06cb774f3ceb480825cc3ebcb717c2979c48ff