More precise size_hint for SmallSet::difference
Summary:
If set `A` has 100 elements, and set `B` has 10 elements, then `A - B` has between 90 and 100 elements.

If set `A` has 10 elements, and set `B` has 100 elements, then `A - B` has between 0 and 10 elements.

So lower bound for number elements in the difference is `a.len().saturating_sub(a.len())`.

Logic seems to be simple.

However, all of these hashsets:
* [std](https://github.com/rust-lang/rust/blob/194140bef501ad3acb00d57c20fb80ee34aa1d3b/library/std/src/collections/hash/set.rs#L1698)
* [hashbrown](https://github.com/rust-lang/hashbrown/blob/e39d245d31db1a0a2ff2a7dfe29d76279e555dec/src/set.rs#L1874)
* [indexmap](https://github.com/bluss/indexmap/blob/aa4ee74bec3f64f452aa107fecd2c1bd04abe66b/src/set/iter.rs#L207)

put lower bound as zero. This is unexpected for rust developers who are obsessed with microoptimizations. Maybe they know something obvious I don't.

[BTreeMap](https://github.com/rust-lang/rust/blob/194140bef501ad3acb00d57c20fb80ee34aa1d3b/library/alloc/src/collections/btree/set.rs#L1657) however sets lower bound to saturating sub.

Reviewed By: ndmitchell

Differential Revision: D40460304

fbshipit-source-id: 089847073170179fa36c935b3fbd2491fd3e78ad