buck2: ConfiguredTargetNode: don't ignore dep data in comparisons
Summary:
In ConfiguredTargetNode, we store deps as ConfiguredTargetNode references for
use in query resolution in analysis (see D37335713 (https://github.com/facebookincubator/buck2/commit/9c9c1b633d70114bdeecbddf986677e4a3ad9ca5)).

We stored them in a LabelIndexedSet, which compares equal if all the labels are
equal.

This means that if 2 nodes differ in a way that only their dep data (but node
labels) are different, we consider them equal (and therefore we make the node
eligible for reuse in DICE).

Unfortunately, that's not quite right. In particular, if you've got A depending
on B (via a query), and B depending on C, and C depends on either Da or Db via
a select, then if the select branch changes, the new B and the old B will
compare equal (because their only data is a dependency on C and that has
not changed), and then we will reuse the node A.

This means that when we resolve queries on A, we will reuse the instance of B,
which will lead us to the wrong C and D. We only access the labels through
the traversal, so then we will query the nodes via DICE, and either get the wrong
data entirely, or possibly an error if the configuration mismatches.

To fix this, this replaces the LabelIndexedSet with a dedicated dependency
container that allows 3 operations:

- Iter, behaves unchanged.
- Hash, behaves unchanged.
- PartialEq now does a ptr eq on all the deps.

Since the deps that a node receives come from DICE, there's only ever 1 valid
value for each ConfiguredTargetNode (the one that's current in DICE), so the
ptr eq avoids doing a deep comparison (which is not practical because the nodes
reference other nodes).

This approach fixes the issue because if the deps have changed in value, then
we will not reuse the node anymore.

You might be surprised we don't hit this more often. The reason is you do need
very specific conditions to trigger it. In particular, you need to have a query
macro, since the query macro is the only thing that looks at the data of the
nodes: it uses it to traverse the graph, and you need to have a node whose
dependency graph can change without its dependents receiving any other changes.

Note that Hash might still be somewhat buggy here. In particular, the node hash
will be unchanged across the scenario described above. This hash collision may
not be desirable. The only way to fix this without making node hashing O(lots)
is to precompute a node hash for each ConfiguredTargetNode and store it.

This happens in the real world: T133069783

Reviewed By: ndmitchell

Differential Revision: D40509285

fbshipit-source-id: ac6b35dae2880dae47593c49a38eaa5ab82a7f4b