buck2: FileDigest: make fields private
Summary:
I want to stop having ActionDigest be a wrapper over FileDigest, because
fundamentally those are different things.

In particular, when we have a TTL on ActionDigest, it's not clear if that is
the action cache TTL, or the TTL of the blob that contains the action.

To fix this, I'm going to give them different types, but that's easier to do
if the external API doesn't expose the struct fields.

Reviewed By: stepancheg

Differential Revision: D40383123

fbshipit-source-id: 329abecc10467d451883a0cc946d22b300f500fa