buck2: replace a panic with a fallible call
Summary:
There is this behavior in DICE where it might try to evaluate nodes that
shouldn't exist in the dep graph anymore to see if they changed.

It'll ignore errors, but not panics. If we start the daemon with exec platforms
then turn them off, if any key that calls `get_execution_platforms_non_empty`
is recomputed for dep checking, we'll likely panic.

The expectations for this not to panic are *not* guaranteed within this DICE key itself,
so it can't legitimately unwrap until this particular behavior in DICE is fixed.

Reviewed By: ndmitchell

Differential Revision: D40523172

fbshipit-source-id: 4c7c4449bf1bc20cfdfc3e53a57b781ee4a92fef