Require Send/Sync on alloc_simple
Summary: Otherwise when we freeze the type, it might get moved around, and we wouldn't notice. I think these constraints used to be present because they are on StarlarkAny, but currently don't need to be (which would be a bug, since if the thing isn't Send, it can be turned into a FrozenValue and things go wrong).

Reviewed By: lmvasquezg

Differential Revision: D40512742

fbshipit-source-id: 816df7d05df69fae32de1f7883d7db1fbac9940f