Don't provide archives as shared libraries
Summary:
Rust libraries which set `preferred_linkage = "static"` don't produce
shared libraries.  This diff fixes a bug where we'd present the static
archive as a shared lib, which leads to various downstream issues
like adding a useless static library to the shared lib symlink tree, and
naming it as a `.a`.

This was also breaking native python builds, as `make_par.py` dies
when it sees an unexpected static lib among shared libs (e.g. P539998514).

Reviewed By: krallin

Differential Revision: D40536126

fbshipit-source-id: 3ad6add4a81eae72c51e32e284591597a427bab2