rust_protobuf_library up to custom_rule (#28)
Summary:
This defines an open source version of `rust_protobuf_library` rule up to the point where it requires `custom_rule`.
The code has been tested on Linux to build `buck2_data` with a `genrule` based replacement of `custom_rule` that is commented out in this PR.

Protoc and the protobuf standard library is fetched in a user defined rule from the protobuf release on Github.

This PR also unifies the required rule shims in `shim/shims.bzl` and defines re-exports to expose them under the appropriate labels.
Relatedly it defines extra utilities to replace internal labels in `deps`, `os_deps`, and `build_env` attributes.

Pull Request resolved: https://github.com/facebookincubator/buck2/pull/28

Reviewed By: blackm00n, lmvasquezg

Differential Revision: D40509774

Pulled By: ndmitchell

fbshipit-source-id: 3e21b6eda3b178a9df637c6df9579b87722384ee