buck2: windows: lookup exec platform not target platform when executing scripts
Summary:
We currently use `_target_os_type` in genrule to decide whether to produce a shell or bat script.

However, that's incorrect, because if we are building a product for Linux, on a Windows host, then
the target platform is Linux but the exec platform is Windows, so we should write a bat script!

This diff updates our OS lookup to be behind a dep so that we can use `exec_dep` for it. This
means we can now read the exec OS type in genrule and run the right thing.

This also updates other places that did similar lookups to either look at the exec platform
if they are informing an `actions.run`, or the target platform, if they are using the platform to produce
an output.

There is one place that's a little awkward: `command_alias`, because it needs both:

- The alias is a script, which needs to match the target platform
- The alias is produced by running either sed or a bat script, which needs to match the exec platform

We can easily get both things. But what if they don't match? There are two solutions:

- Ensure that doesn't happen (this diff does 50% of this) by adding a constraint such that if the target
  platform is X, only X is eligible as the exec platform.
- Be compatible with both by running `sed` on Unix and `find_and_replace.bat` on Windows regardless of
  whether we are doing so on a shell script or a batch script.

For now I'd like to do the former because i don't think we urgently need to fix this and because the latter
isn't a general solution (though here it'd be fine), but I ran into an issue with exec platform resolution, so
I'm going to try to debug this.

Reviewed By: themarwhal

Differential Revision: D41181560

fbshipit-source-id: 6e27c8c026e2bbe7a2d2ae143d54b2afe6dbb4f0