make `with_dice_ctx` not take ownership of ServerCommandContext
Summary:
As the title says. This is used in the next diff to fix an issue with LSP.

Looking at the code, `with_dice_ctx` doesn't _really_ need to take ownership of the `ServerCommandContext`. I chatted with bobyangyf about adding another method (e.g. `with_dice_ctx_ref()` or something) that doesn't take ownership, but he recommended just changing the original API if possible. An alternative would have been for the function passed into `with_dice_ctx` to return the `ServerCommandContext` that it was called on, but I think this is a better approach, even though it does touch a lot of files.

Reviewed By: ndmitchell, bobyangyf

Differential Revision: D41088724

fbshipit-source-id: 58156c6e97ef58f604ffb1d35ea605178f4fb310