Retain name when symlinking prebuilt_jar
Summary: We symlink the prebuilt_jar so that we have an "owner" for the `.jar` artifact. This changes the symlinked jar's name to be the same as the name of the original jar. That's needed to keep some annotation processors happy - e.g. if they look for the name of the jar, now they will find it as expected.

Reviewed By: navidqar

Differential Revision: D41268252

fbshipit-source-id: 895e70dd7e97d2da6501be137bb38d273285b429