Adjust stats so they can be used incrementally
Summary:
Previously the stats for file notifications wanted you to know how many notifications there were in advance, and printed out messages as the stats were being accumulated. That works fine if you don't process any events until the sync point, but if you want to do it incrementally in the background, it doesn't work. Change it to buffer everything.

Slight regression in that we print things after invalidating DICE rather than before, so might give the details slightly later. But fixed in the next diff.

Reviewed By: krallin

Differential Revision: D41301384

fbshipit-source-id: baaeab959f2324fe3ef9af1ede933a93d9abfeb5