Fix Python tools used in Python toolchain to use python_bootstrap_binary
Summary:
On windows, python files need to be explicitly run with the python interpreter. While running the python file directly works on unix, we cannot assume python files are runnable on its own.

To explicitly specify certain files to be run with the python interpreter, use the `python_bootstrap_binary`.

#forcetdhashing

Reviewed By: ndmitchell

Differential Revision: D41267039

fbshipit-source-id: 9a55e7c2acad18426aea78d9b92253133d9f4b88