Default to pika for unrecognized platforms in the bxl script
Summary:
When using buck2, we were falling back to the default platform when working in fbobjc. This is because the bxl script was returning an unrecognized platform/toolchain pair "xcode" and "iphonesimulator-x86_64". This changes the bxl script to return "pika" if the platform doesn't include "platform" (platform009, platform010, etc.).

There may be issues on Windows platforms later on when they migrate to buck2 based on this, but those are hard to anticipate right now. One alternative would be to add a map of platforms that translate to pika: iphonesimulator-x86_64, iphone-arm64, macos-arm64, etc.

Differential Revision: D41354057

fbshipit-source-id: b292e69d418c2dbcd463be75a6a0aeb9e75459cb