Don't use Duration::from
Summary:
You used to be able to convert between prost Duration and std Duration with into/from. Since prost 0.11 (the one after we are on, but one we will need to upgrade to eventually) that conversion has gone, so only use try_from, which is available on both.

Two strategies here - if we weren't in a fallible block, and were generating proto, I went with a None, otherwise I bubbled up the error. In practice, this only happens if you have a negative duration, which I don't think we ever do, so should be fine.

Differential Revision: D41497042

fbshipit-source-id: 56231f3be9f6911b5780b7a7c05066072dbe6baf