buck2: rename `add_arg_string`
Summary:
I want to allow exposing lower-level APIs to write directly to the buffer
to reduce allocations when we produce command lines. But, to do so, we need
to be able to indicate "I'm appending" vs. "I pushed an element".

I'll add those APIs in a later diff, but for now this diff updates the existing
function to clarify what it does (add an element) and not how it does it (by
receiving a String).

Reviewed By: ndmitchell

Differential Revision: D41493664

fbshipit-source-id: a3e8011dd8bf83f42c718236ca737c0386e60a7e