Upgrade the rust image
Summary: We don't actually rely on the Rust image for getting Rust (beyond rustup), but we do rely on it for our Ubuntu version, and thus our protoc version. The 1.49 we used to have gave us protoc 3.6 which didn't support optional, whereas the 1.65 has upgraded Ubuntu enough to give us 3.12 with protoc.

Reviewed By: themarwhal

Differential Revision: D41681644

fbshipit-source-id: 77b3f16f2067af110dae759dfd6d1627a7d2ffac