buck2: project fs: set executable for all users
Summary:
When we write an executable file we currently set only the executable bit for
the owner. However, when downloading files from RE, "executable" files have
their executable bit set for all users. This is inconsistent, and a little
confusing, because "executable" now means different things depending on
how the output is materialized.

For the sake of consistency (and also usability), set the executable bit for
all users.

Reviewed By: jjmaestro

Differential Revision: D41763950

fbshipit-source-id: 276172b18e6caece36a8ee0a0c49f87c7b142dd8