Improve error message on unset edition
Summary:
Previously this would pass `--edition=None` to rustc, which doesn't make a lot of sense.

In the case that neither `ctx.attrs.edition` nor `toolchain_info.default_edition` is set, I definitely think buck2 should fail the build instead of not passing an `--edition` flag to rustc. Rustc's default edition is (and seems like will always be?) 2015, which is not a sensible default in 2023+. It's just an accident of rustc's history that it is forced to behave that way.

Reviewed By: zertosh

Differential Revision: D41778098

fbshipit-source-id: 709eaccabd173d630be2c6746be0fbe4cb82d07c