buck2: `target --resolve-alias`: verify that the aliases exist
Summary:
Like it says in the title. Right now we happily just tell the user how we'd
evaluate an alias, but we don't check that whatever it evaluates to is correct.
And, since on the face of it `foobar` can either be an alias called `foobar` or
just `$CURRENT_CELL//$CURRENT_PACKAGE/foobar:foobar`, that means we never error
out.

In contrast, Buck1 does, so let's add this validation and be consistent. Note
that we don't actually tell the user if the problem is the alias doesn't exist
vs. what it points to is broken, but the error messages should make it sort of
clear (because tell you what we looked at when we encountered the error).

Differential Revision: D41798396

fbshipit-source-id: b2868cc266859267d1f2066fe48f1f4fe56bf3f2