Improve how buildifier is configured in reindeer.toml
Summary:
The current way `buildifier_path` works is not what most users of buck2 outside of Meta are going to want -- for example https://github.com/dtolnay/cxx. Other than massive monorepos it's not typical to have a `buildifier` executable vendored into the repo to be accessed at a relative path relative to the reindeer.toml. Users need to be able to install `buildifier` and make reindeer run it from their PATH.

This diff replaces `buildifier_path` in reindeer.toml with a `buildifier = true` setting that makes reindeer run "buildifier", not relative to the reindeer.toml, just from PATH.

We keep a way to set a relative path though.

Reviewed By: zertosh, diliop

Differential Revision: D41794618

fbshipit-source-id: ef1650d96325d20c997338859922c4812e6cdf1b