Make root parameter in group optional
Summary:
This allows the root parameter of the group tuple, which specifies the
node for which to apply the mapping, to be `None`, in which case the
deps of the top-level node/executable will be used.

This essentially ports the existing semantics from `cpp_link_groups.bzl`
into the prelude (and out of the macros).

Reviewed By: BrandonTheBuilder

Differential Revision: D41723617

fbshipit-source-id: f0d9c1406fdfc3014b93d3d822c968b9b993cb18