buck2: materializer: deferred: make sqlite calls synchronous
Summary:
Read here for why I'm doing this stack and for benchmarks: D41774542

 ---

Like it says in the title. Make sure you read the linked diff above to understand
why this is OK;

Note that there is a mutex in there in SQLite to make it `Sync` but in reality
there is only one thread acquiring it.

Reviewed By: scottcao

Differential Revision: D41774551

fbshipit-source-id: 758dc562abd8d812c20257d482fff8b34c13386d