buck2: materializer: deferred: do SQLite insert in materialization_finished
Summary:
Read here for why I'm doing this stack and for benchmarks: D41774542

 ---

Like it says in the title. The current model is hard to reason about because
half the state updates happen in the materializer state (the ones we use at
runtime) and the other half happens in individual requests (the ones that go
to SQLite).

This also leads to some not-obvious bugs, like futures blocked everywhere
waiting on SQLite when it tries to do a `sync` which went unnoticed until now
(it went unnoticed because we rolled this out to MacOS where sync is a noop).

Note that this defers updating the SQLite status a little bit. Prior to this
we would do it as soon as the entry is materialized, and now we do it after
the deps are materialized. In practice, that's unlikely to matter since we
only care about this state across restarts and you don't typically restart
between writing a thing and writing its deps.

Also note that this means we are blocking the command thread on SQLite, but see
later in this stack for benchmarks showing this does not regress performance
(in fact it might improve it).

Note that this diff introduces a soft error for failing to write to SQLite,
which seems quite wrong, but until now we had no centralized handling of
those errors, so that's why we have to have this.

Reviewed By: scottcao

Differential Revision: D41774547

fbshipit-source-id: c9bc62209241d0890919d146d5fcba3de9761f36