buck2: materializer: deferred: move SQLite invalidation out of clean up future
Summary:
Read here for why I'm doing this stack and for benchmarks: D41774542

 ---

Like it says in the title. This was put there because we had to defer it until
after the materialization may have written its state, but after the previous
diff, we only write the state when materialization finishes (which is gated on
a version check), so this is no longer needed.

Also see the previous diff for context.

Reviewed By: scottcao

Differential Revision: D41774556

fbshipit-source-id: 94791566897b137fd6e46a9a0fa68f79a0e1f17c