buck2: when `defer_write_actions` is enabled, kick off writes in the background
Summary:
This is the diff this whole stack has been about enabling. Let's recap how we
got there:

We would like to put writes into through the deferred materializer for a few
reasons:

The first is state keeping, which means that we can track & reuse existing
files. Writing files to disk (and, in fact, deleting existing files on disk)
is where we spend most time in writes (as opposed to generating files),
so skipping the I/O is a great thing. It also is a requirement for us to
be able to implement `buck2 clean --stale`, which requires us to know
what files exist.

The second is putting writes off the critical path. We don't actually need
files to be on-disk to proceed with a build, we just need their hash. We can
hand off the writes to the deferred materializer and proceed with the build.

This also has nice properties in terms of reducing our concurrency for writes
(which we've seen in the past yileds better I/O performance than throwing
millions of requests at the FS) without forcing too many task context switches.
Finally, it allows us to couple cleanup & write, which again means fewer task
context switches.

Now that we've gone over why, let's see the how.

The initial way (6 months ago) I had POCed this was to just keep the files in
memory forever.

The downside is that this regresses memory quite a bit (because we have a
non-trivial amount of files). It also still causes us to do some I/O anyway to
clean the output path (lest we leave a potentially stale file there and that is
confusing). Besides, it didn't actually enter those files into the materializer
SQLite state.

A better way is to do the write and call materialize immediately (basically, do
a `declare + materialize` call). This is doable, but not good for perf. We
spawn 2 Tokio tasks (one cleanup, one materialization) for every single file
we write, and we write a lot of files. This improves memory at rest (we don't
keep the files in memory) but regresses memory at peak (very) substantially
(by about 20% on the benchmark below).

So, this stack is about doing this in a better way, which is by still kicking
off the writes but spawning no Tokio tasks at all. To do this, we need to:

- Not require a Tokio task for deleting from the SQLite state. This means we
  need to do the SQLite delete operations in the command thread. As
  benchmarked, this is not regressing perf after enabling WAL + disabling
  fsyncs.
- Not requiring a Tokio task for the write either. This means moving the insert
  to the command thread as well, which then allows us to just send off a
  `IoRequest` to the blocking executor, which signals the deferred materializer
  when it's done.

This diff does this latter bit, which is possible now.

See benchmarks below in the Test Plan.

Reviewed By: scottcao

Differential Revision: D41774542

fbshipit-source-id: b68d113aaa2e6e8de637425a0f9f5b38f1409721