Support using a target_label as cmd arg
Summary:
In the Rust rules in the prelude, we are interested in passing `ctx.label.raw_target()` as a command argument.

https://www.internalfb.com/code/fbsource/[4df447547768ee7f530e9871f89d4feb3ab7ca35]/fbcode/buck2/prelude/rust/build.bzl?lines=123%2C135-137

It has a `ToString` impl so `str()` works. But generally I don't have a good intuition around what values can and cannot be directly used with `cmd_args`. A `target_label` is conceptually string-like enough that it made sense to me to try passing one directly. In my mind it's a string + just some accessors for package, name, cell. Unlike for example `configured_target_label`, which is not straightforwardly understood as just a string.

It would be interesting to try to articulate in the documentation of `trait CommandLineArgLike` some principles for what should implement it. For example are integers a good fit? I have no idea. `cmd_args("--timeout-seconds", 60 * 45)`

Reviewed By: krallin

Differential Revision: D41779871

fbshipit-source-id: 0a21bd684c11eb3eac695aaa689896344f1cb2c5