Don't force load archive from private link group deps
Summary:
Don't force load an archive unless it's a public link group node. This does not get rid of
redudant external symbols because the symbol will still be loaded if it's referenced
internally, but this should help get rid of duplicate symbol errors from when the same
source file is included in multiple targets.

Reviewed By: andrewjcg

Differential Revision: D41625595

fbshipit-source-id: 6aeccd02371a1bbb6d08dbb96ecd81e863ffee37