buck2: dep files: soft-error if the file cannot be read
Summary:
Like it says in the title. I'd like to stop pruning dep file state on Watchman
restart (because it can help us mitigate the impact of Watchman restart!), but
the reason we had added dropping state on Watchman restart was because of
spurious errors reading dep files.

For now, let's tolerate the individual errors instead via a soft error.

Reviewed By: ndmitchell

Differential Revision: D41994472

fbshipit-source-id: 792c52cd41440ee2474739ef8bdd1bcd08096bce