Add build_and_connect
Summary: In open source Buck2 we want to connect to the GRPC client eagerly, since given how Buck2 works, there is no point in delaying it. Introduce a separate function to do that, which must be in async (so you can spawn the connections) and Error (so you can fail).

Reviewed By: krallin

Differential Revision: D41838601

fbshipit-source-id: 0d0c57191ad2ef76ed88ec08e173986dda0754ec