Add --tracked-only flag to clean --stale
Summary:
Adding previous implementation which does not scan buck-out behind a flag.

Only cleaning tracked files has the nice property that it is extremely fast when there is not much to clean, may be convenient to run on a regular basis on laptops if full `clean --stale` can't be scheduled, or if any untracked artifacts are being deleted by `clean --stale` that shouldn't be.

Reviewed By: krallin

Differential Revision: D41866801

fbshipit-source-id: 768aec7e66f7cbe208bddf2ae713511f8580f059