buck2: when the daemon is killed, tell the user as much
Summary:
Like it says in the title. If someone runs `buck2 kill`, then notify all
waiting clients that this happened, and if those clients later experience
their stream breaking (because the daemon was killed before it finished
handling the request), tell the user that this is what happened.

There's a bit of complexity in making this work reliably. In particular, we
want to ensure that this message gets ahead of the queue. To achieve this, we
send it through a separate channel (instead of our regular event stream).

See the test plan for how much more informative this output is now.

Reviewed By: ndmitchell

Differential Revision: D42033377

fbshipit-source-id: b3da04650db15ac9f8cf9a3aa05eb63f5df4cd4d