buck2: data: add explicit presence for exit_code
Summary:
An int defaults to 0 when not set in protobuf, which turns "didn't have
an exit code" into "all is well". Fix it.

In practice, we only use the exit code for presentational purposes and logging
so it's not a huge issue, but it's easy to fix by just using explicit presence
for this field, so let's do that.

Note that explicit presence does not change the wire format but only the
codegen, so this is not a breaking protobuf change.

Reviewed By: ndmitchell

Differential Revision: D42066078

fbshipit-source-id: d02495943a56f7cc423fd312bd55fd432b6e433b