Use fnv as a hasher everywhere
Summary:
This switches all `HashMap`s and `HashSet`s in dice to use the fnv hasher. Fnv is much faster on small inputs, which the keys of these structures typically are.

Unfortunately, I did measurements on a bunch of targets and was not able to find a statitically significant improvement from this. This is probably still a principled thing to have in anyway, but would also understand not wanting to take this.

Reviewed By: bobyangyf

Differential Revision: D41976481

fbshipit-source-id: f65b2799502447ce2d017098360861e9edfcbe09