Make sorted require reverse to be a boolean
Summary: The Starlark spec merely says that True should reverse sort. It doesn't actually say that False should normally sort (although you can guess that), and doesn't say what happens if you pass a string there. In Python3, it requires either a bool or an int, but you have to assume that's weird historical reasons. So suggest we interpret the parameter as must be a boolean, which is both plausible and what most people expect.

Reviewed By: stepancheg, bobyangyf

Differential Revision: D42284174

fbshipit-source-id: e9ef68df38d2b1eb4640d0bd86cb578dca68db8a