buck2: include defer_write_actions status in materializer state versions
Summary:
Like it says in the title.

This matters because if you build with defer_write_actions then later build
without it, you'll get a bunch of warnings about artifacts getting materialized
without being having been declared first, like so:

```
2023-01-02T10:39:35.482772Z  WARN materialize_artifact{event_dispatcher=EventDispatcher(TraceId = 46b66c5d-8162-4cb2-a8aa-a0564b70b53d) path=buck-out/v2/gen/fbcode/9dbbb55f355c29b9/tools/build/buck/wrappers/__ld__/ld}: buck2_execute_impl::materializers::deferred: Expected artifact to be marked active by declare path=buck-out/v2/gen/fbcode/9dbbb55f355c29b9/tools/build/buck/wrappers/__ld__/ld
```

I don't actually think this would break something, but the logspew is certainly
not OK, so let's get rid of it.

Reviewed By: scottcao

Differential Revision: D42305919

fbshipit-source-id: c1abf6cfd2ba9b43b4ca88d7340f05d6721c7b6b