buck2: soft error: broadcast soft errors to all connected clients
Summary:
We currently don't capture those in event logs at all, which is not great
because it means that if we get a bug report where a soft error might have been
relevant we have zero visibility into it.

So let's fix that by sending them.

Later in this stack, I make this a bit precise when possible.

Reviewed By: stepancheg

Differential Revision: D42450275

fbshipit-source-id: aed5ed58724b07f75bb958977359c3730c4dbd91