buck2: log when people use console interaction
Summary:
This feature has had a fair amount of negative push back. Initially some of it
was about bugs, but we also have feedback about personal preference on how
it behaves that isn't really "fixable" in code:

https://fb.workplace.com/groups/buck2users/permalink/3337507699838921/

So, let's work out if it's actually used. I suspect it's mostly useful for
debugging (i.e. by the Buck team, or perhaps just me), so assuming that's
indeed the case, I'll just flip the default.

Reviewed By: stepancheg

Differential Revision: D42496229

fbshipit-source-id: 37a37bf7054f4dd381a63633985bdeadf52181e8