buck2: materializer: when reporting CAS download errors, include the RE message
Summary:
Like it says in the title. Right now we tell you the digest origin, which is
useful to tell if the error is on the Buck2 side or the RE side, but we don't
capture the error from RE, which makes it not very nice to debug for the RE
folks when it's on their side.

Reviewed By: johansglock

Differential Revision: D42495349

fbshipit-source-id: 1b4ece760da123079ff8457d7aa5d7332259d7ac