Make rustdoc testable: `buck test :target[doc]`
Summary:
This diff implements the Buck2 version of `cargo test --doc`, which:

1. performs macro expansion,
2. extracts all example code from Rust documentation,
3. wraps it in a standard template along the lines of `#![allow(dead_code)] fn main() {`…`}`,
4. compiles the resulting code if not marked ` ```ignore`,
5. and runs it if not marked ` ```no_run`.

### Future work

Rustdoc does not have an implementation of rustc's `--remap-path-prefix` flag, so if a piece of example code does not compile, the paths in the resulting diagnostic are to buck-out instead of the original source file. You can see this in the test plan of the diffs in the rest of the stack. I am interested in contributing a remap-path-prefix implementation to rustdoc.

Separately we might need a `doctests = False` opt-out attribute, much like the existing `unittests = False`.

Reviewed By: zertosh, diliop

Differential Revision: D42512484

fbshipit-source-id: c7eca48baeee38dc714e0b5fe4d1cbc2af562397