doctests=False for targets with currently failing doctests
Summary:
Here is a test run of all Rust doctests currently in fbcode. 4074 passed, 272 failed. This codemod applies `doctests = False  # FIXME` to all the `rust_library` targets with currently failing doctests; then we can incrementally enable as the tests get fixed.

https://www.internalfb.com/intern/testinfra/testconsole/testrun/4222124797890645/

The rest of this stack does similar for various other rule types, like `rust_python_extension` and `rust_protobuf_library` and `smart_service_rust`.

Here is a test run of all remaining enabled doctests at the top of this stack:

https://www.internalfb.com/intern/testinfra/testconsole/testrun/5066549726206406/

To fix a doctest, remove the `doctests = False # FIXME` and do `buck2 test //path/to:thetarget`. All example code in the documentation of the target will be compiled and run. You need to fix the example code to make it work, adding `doc_deps` to the target if there are other crates needed in the documentation which aren't already `deps` of the library.

Reviewed By: zertosh

Differential Revision: D42565364

fbshipit-source-id: a9ea3939d50ac225134f868d0ab44c1272bfd116