fix dsym materialization when using test command
Summary:
When running `buck2 test ...` only artifacts from `ExternalRunnerTestInfo` are materialised, namely those in `command` and `env` fields. I want dSYM to be materialized too when executing:
```
`buck2 test :foo[dwarf-and-dsym]`
```
One option is to add dSYM as `hidden` to `ExternalRunnerTestInfo.command` but it's ignored by tpx. So add it as hidden to `XCTEST_BUNDLE` env variable.

Reviewed By: krallin

Differential Revision: D42585404

fbshipit-source-id: 7ace383d59e1b50c5a61081aea4991a18c6b5495