Default to no message batch size
Summary:
In D42178265, backoff was implemented in buck2's scribe client to throttle our rate of sending messages. The thought was that this was why we consistently saw so many missing messages in corp.

Unfortunately, a low batch size (3k) was picked as the default. While this is fine for corp (~6k msgs/sec), it's paltry for prod, where we routinely send up to 5x that during a normal fbcode build.

This diff removes this limit but allows re-instating a batch size via the optional `buck2.event_log_message_batch_size` parameter. When set, this is what buck2 will use as the new message batch size limit when pulling messages off its internal scribe queue.

When landed, this diff brings us back to the state we were in prior to landing D42178265 (e.g. draining the entire queue of messages in one batch send), except with some smarter backoff logic if we get `WriteMessageResult.TRY_AGAIN` from scribe/scribble.

Reviewed By: krallin

Differential Revision: D42655315

fbshipit-source-id: 2170738f88581a12e6b586000fc0261698035fd0