ThinArcStr
Summary:
It is like `ArcStr`, but allocates the size next to string data instead of in value.

This saves memory. Consider this example: arc str of length 20 with 1000 copies.

`ArcStr` would take 1000 * 16 + 20 = 16020 = bytes.

`ThinArcStr` would take 1000 * 8 + 20 + 4 = 8024 bytes.

Also it will make enum packing more efficient, for example, struct like `Option<ThinArcStr>` takes 8 bytes instead of 16 for `Option<ArcStr>`, and this is important when `Option` is `None`.

This is used in the following diff D42651347.

Reviewed By: JakobDegen

Differential Revision: D42651348

fbshipit-source-id: 13ae335517c0bcce60e32c1326e8a22d21f9d6cc