{Coerced,Configured}AttrFull
Summary:
`CoercedAttrFull` and `ConfiguredAttrFull` contain:
* name
* attribute type (currently unused)
* attribute value (`CoercedAttr` or `ConfiguredAttr`)

Started doing to be able to remove dep type from `AttrLiteral::Dep`: this information is redundant, it is available in the attribute type. But this is larger change.

For now it is useful as is because for example `ConfiguredAttrFull::traverse` function wrapper adds context: the name of attribute.

Reviewed By: bobyangyf

Differential Revision: D42666959

fbshipit-source-id: adab507e32a45121a04f7d8a0efff6d34c1b3977