Add core support and config flag for execution platform fallback behavior
Summary:
Currently when we don't find a matching execution platform we just use
the "unspecified_exec" platform. This was convenient for migration, but we'd rather
make this an error.

This adds the core support for user-defined fallback behavior that can be one of:

1. a full ExecutionPlatformInfo to use that won't be checked for execution
compatibility. Note though that the exec_deps will be configured with it and so
will be marked compatible/incompatible as appropriate (but exec_compatible_with
is also ignored and wouldn't show up in anything as compatible/incompatible).
2. error, where we'll give an error at the point that we fail to find an execution platform
3. use_unspecified, just explicitly requests the previous behavior

If none are specified we'll continue the previous behavior (for now).

Even before we move the default, there's some use for this config flag. Right now, we can
run into an issue where we hit the fallback case and then in some dependency down in the
graph we hit an error due to trying to access the configuration constraints (like trying
to resolve a select). In that scenario, we currently can't even use
`buck2 audit execution-platform-resolution` to understand why we hit the fallback for a
node because we'll hit that same error. With this change, you could re-run the cquery
with `-c build.execution_platform_fallback=error` to see the full explanation for all
the skipped execution platforms.

It might be nice to also add an option for "incompatible" such that any target that
can't find an execution platform is marked incompatible just like target compatibility
(and like target compatibility there's risk that that leads to things silently changing
from working to being skipped when an error is introduced).

Adds the macro/target-level bits to make this configurable in buckconfig.

Reviewed By: krallin

Differential Revision: D37541562

fbshipit-source-id: 46f871cbe1ed69bb1f61cb916db18272fbbea9aa