update BuckdClient::kill() to always forcibly kill the daemon
Summary:
Previously `BuckdClient::kill()` would fail with an error if buckd returned an
error from the graceful kill request.  This caused confusing behavior for
users when the buck daemon got into a bad state, and left them with no easy
recovery mechanism: connect errors would suggest running `buck2 clean`, but
this attempts a graceful kill first and would fail.  `buck2 kill` also failed
in a similar way.

Notably this error can occur when the client cannot authenticated to the
server successful.

This diff updates the client behavior to always continue and try to terminate
the server process with OS-level functionality, even after an error response
to the graceful kill request.

Reviewed By: krallin

Differential Revision: D42648061

fbshipit-source-id: 6f2f10aa3df14fb1605f1287e5c295a2469af17b