buck2: batch events we send
Summary:
It is *a lot* faster to send larger messages to Tonic than small ones. This
comes down to:

- There is message passing overhead for every single message all the way from
  Tonic to the h2 crate's layer. You do less of that, you get less overhead.
- The h2 crate will flush messages if they are 256 BYTES or more. Most of our
  messages are more than 256 bytes, but they're still pretty small relative to
  a max TCP packet size to localhost. If you send messages 1 by 1, you spend
  all your time in the TCP stack. If you send batched messages, you don't.

Note that this only batches messages that are ready, so it doesn't introduce
latency. It only kicks in if you've got a queue to send.

This is super easy to do, and pretty impactful and saves us a GB of two at
peak and reduces our event latency..

Reviewed By: blackm00n

Differential Revision: D42781413

fbshipit-source-id: 361e3eed320b97d4e753277db3c2d480ea061e90