buck2: materializer: immediate: add a critical section
Summary:
This seems a bit off. If this gets cancelled then Buck will just forget this
ever existed and we might start materializing multiple times into the same
location.

This might be the cause of [this][1] but it's a bit hard to say conclusively.
That said, this looks like the only place in the immediate materializer where
we're missing a critical section so this sounds like a good place to start?

This is (unfortunately) another (great) example of why *just dropping stuff* isn't a
good idea at all and we should just stop doing that, but one thing at a time.

[1]: https://fb.workplace.com/groups/buck2users/permalink/3348342228755468/

Reviewed By: milend

Differential Revision: D42800838

fbshipit-source-id: 3c2d8d7ed61f6519034823d1d12039e33eecb971