buck2_event_observer: move IoState in
Summary:
Like it says in the title. Actually putting this in EventObserver comes in the
next diff as this was getting a little big;

Note that as usual, the UI component says behind, and only the state moves, but
sometimes those things are a bit coupled and for now I've not bothered with
cleanup.

Differential Revision: D42782389

fbshipit-source-id: 0941be7b6b07cda63493a414d75e4123eb2b6fe0