Shuffle around of Introduction content
Summary:
These changes were done after conversation with Wendy Yu and Brian Johnson, and a Workplace chat with Neil Mitchell.

NOTE: Lines 6 - 16 are intended to be a conversation starter on what is hoped will become an introduction to the question "What is Buck2" WITHOUT talking about or referring to Buck1.

=========================================================
The RFC information has been moved to its own page - see D42806516

There will be future updates to the Introduction page and the Getting Started page.

Differential Revision: D42843751

fbshipit-source-id: b617965a103231072f5e2b310c23fdf64f70a5dc