Convert some results into anyhow::Result
Summary:
`SharedResult` is needed in dice, but it should not be used in interfaces, because it is implementation detail of specific operations.

In particular, there's no need to convert regular errors into `SharedError`. Using `SharedResult` may lead to misunderstanding that each operation returning `SharedResult` has dice calculation behind, which is not always the case.

Reviewed By: bobyangyf

Differential Revision: D42832032

fbshipit-source-id: 5d56cdc890c4c00d23211ea71bd6e3250d9c4d4d