buck2: CancellableFuture: allow entering critical sections after being cancelled
Summary:
Like it says in the title (note: this is how the previous DropCancel
implementation worked).

If something enters a critical section but all callers are gone, we need to
either error out (but that is not good because it might have observable side
effects like events being fired), or actually carry out the execution.

This does that.

Reviewed By: ndmitchell

Differential Revision: D42852567

fbshipit-source-id: 39c62d275ce5926b29d86bc2675f2cf3ec576660