buck2: dice: don't apply state updates from cancelled executions
Summary:
If an execution was cancelled, we shouldn't apply state updates, because the
value we get might be the result of the cancellation we observed.

There's two ways we could implement this:

The risky way would be that when we exit any section that might have blocked
cancellation, we check if we've been cancelled, and if we have, we do
`futures::future::pending().await;`.

In theory, that would mean control returns to `CancellableFuture`, and then
`CancellableFuture` would go "ah, we are pending + we are cancelled, return
`None`".

The problem with this is that if it ever goes wrong, the symptom is a future
that is sitting idle somewhere, and that isn't great.

The safer way is what this diff does, which is that we explicitly check for
cancellation and then omit the updates if so, though it does come at the cost
of exposing the possibility that a cancelled future might return a value in the
`DiceError` type, which is a little awkward, but also not the end of the world.

For motivation, see: D42804635

Reviewed By: bobyangyf

Differential Revision: D42881737

fbshipit-source-id: d86498bac057359d176c07e8ab7d6de42b4343f4