ProjectRoot::relativize_any
Summary:
Utility to relativize a path which is not normalized.

And not just normalize it, but also canonicalize it.

So if absolute path is passed using different symlink (e.g. `~/fbcode/xxx` where repo root is `~/fbsource`), it should still relativize it correctly (as `fbcode/xxx`).

Used in the following diff D42864459.

Reviewed By: bobyangyf

Differential Revision: D42864460

fbshipit-source-id: 179c40f9e4233a908e154c4fd23b7b3ccca42c1e