buck2: cancellable_future: check some assumptions + don't exit if never entered
Summary:
We assume that the things created by `critical_section` and
`with_structured_cancellation` are dropped by the CancellableFuture that
created them, and if they aren't then, then bad things happen.

This diff makes us check those assumptions. In an ideal world we would express
the idea that "this future can be Send but only if whatever wraps it is also
Send", but that's kinda hard to do so we don't.

We do this by checking the address of the execution context, which doesn't
move because it's behind a Box.

However, we do sometimes (in the materializer) create critical sections off of
a CancellableFuture then drop them on a CancellableFuture, so we allow that
and skip exiting if we never entered.

Reviewed By: ndmitchell

Differential Revision: D42920743

fbshipit-source-id: 8a78fab281a746e458342dbf17b6f52222020e39