buck2: audit exec platform resolution: be less ambiguous
Summary:
We currently show the config of the exec platform, which tends to tell you
which exec platform you're using, but it's not necessarily super obvious.

Update the terminology to use the actual id of the platform and then put the
config next to it for visibility (though audit providers on the exec platform
would give that to you as well).

Reviewed By: milend

Differential Revision: D42919082

fbshipit-source-id: 4d2e953be4fac724ee6a6857025e2d6598b88453