buck2: cancellable future: make it less panicky
Summary:
Like it says in the title. The original idea here was to look at which future
is currently executing in order to exit critical sections and the likes, but 2
failed releases later, it's somewhat self evident that we aren't entirely
capable of getting this right, so let's make this safer.

What this diff does is that even if a critical section is exited by someone
that *isn't* the future that created it, that future will still actually get
out of the critical section.

Note that we don't attempt to wake up the task whose critical section exited,
before critical sections don't interact with wakeups (they just make
cancellation potentially take effect later).

Reviewed By: IanChilds

Differential Revision: D42996482

fbshipit-source-id: d25c46451219399964194a9e82e81b45fd1d8006