Make alias resolution part of the main if block
Summary:
Previously alias resolution was very much a special case, with an early return. Now pull it into a single if.

I do end up creating the formatter in two branches, but that seems clearer code to me as you only have one nesting of if. It's only a few lexemes and no duplicate work, so not a big deal either way.

Reviewed By: IanChilds

Differential Revision: D42997448

fbshipit-source-id: 80d0e6337d48d6e05455b1406e29e27563b605c4