buck2: materializer: deferred: implement iterator manually
Summary:
The current iterator is pretty inefficient due to chaining and dynamic
dispatch. This makes it faster + create new possibilities to optimize (see
later in this stack).

It's a bit of a shame that we have to implement this twice (for into_iter()
and iter(), but you can't be generic over mutability in Rust).

Reviewed By: scottcao

Differential Revision: D42964076

fbshipit-source-id: 5398f69bc51f710e7f7ed737d51366cf827a0abd