kotlincd_jar action also produces a class ABI
Summary:
Per previous diff, this saves us an extra action to produce the class ABI (which is what we generally are going to use for compiling against etc anyway), and is fast to produce. In subsequent diffs we'll be using `jvm-abi-gen` to produce the class ABI from the main library compilation anyway.

For now, this saves about 13k actions for `fb4a` builds (that will go down as Kosabi rolls out though).

Reviewed By: navidqar

Differential Revision: D43045399

fbshipit-source-id: d8704ac80d995f0a543be8a50341e30632eba5c2