Use absolute paths for symlinks in Windows
Summary:
I'm mainly trying to solve two things here:
1. Fix buck2 build error seen on Windows, where we call `canonicalize()` during `symlink` on a target that hasn't materialized yet. This happens very consistently, even if we fix the order or materialization here: https://www.internalfb.com/code/fbsource/[b6bbe6d89f9464a89ca5e58f76c15aac92e7c008]/fbcode/buck2/app/buck2_execute_impl/src/materializers/deferred/mod.rs?lines=1382-1396 .
2. Add support for adding the \\?\ prefix so we can support paths longer than 260 characters.

Previously, some logic that invokes `canonicalize()` was added here D38365802 (https://github.com/facebookincubator/buck2/commit/ca43c51d4410bac4a9dc22857187d43c3e5eb673) to  get round how Windows symlinks are "relative to real path on the filesystem and not the path you're accessing them with". So the logic added was to check if the realpath of the target was where it's seems to be (`canonicalize()`) and use the absolute path if it's not. But this only works if the target path (or its parent, in this case) is always materialized before the symlink. This is not true, because of the way we materialize files (see above) and other reasons (not yet discovered).

As far as I understand, it should be safe to always just use the absolute path so that we avoid this relative path issue. This way, we're not dependent on whether the target file exists on this system, an assumption we clearly cannot make as it is.

To canonicalize the path, we call canonicalize(). If that fails, we find the common ancestor between the symlink and the target (which lies somewhere between symlink's parent and repo) and canonicalize that.

This still leaves a potentially bad bug where we create a file symlink when trying to create a symlink to a directory that doesn't yet exist. This issue exists on master as well. It would be very good to fix this, but we have to understand why the file isn't materialized properly first.

Reviewed By: krallin

Differential Revision: D42741818

fbshipit-source-id: d7123669d798e9d98cc2205d3f1eee853daea13e