Add abi_as_dir to JavaClasspathEntry
Summary:
We want to have per-`.class` file dep-file support for JavaCD and KotlinCD, but we don't have any way to track each `.class` file inside a `.jar`.

Instead, we can have a separate unused input that we pass to the action that is a directory representation of the `.jar` file, and then we'll use each file within the directory representation for our dep-files.

This just adds a field that will store the directory representation, but it is not yet populated anywhere.

Reviewed By: navidqar

Differential Revision: D43087061

fbshipit-source-id: 74d758bf595c449a240f4454c1569f8502af7312