Add command final count value for action fallbacks
Summary:
It was brought to light in [this post](https://fb.workplace.com/groups/930797200910874/permalink/1066676623989597/?comment_id=1067046457285947) that when RE actions fails and falls back to local, there is no warning message that informs users that the local action was a fallback.  This proves painful for debugging, and so we would like to introduce the following UX improvements for prompting on action fallbacks:
1. Show a count of the number of fallback actions (next to local / remote / etc.).
2. Show a field the first time it happens.
3. Show a message at the end.

This diff adapts buck2 to do the first, which is to show a running total of number of actions that went into fallback.

### Determining fallback actions:

Currently it doesn't seem to be a blessed field to label an action as a fallback action, and so we introduced a function to determine that based on a command's history.

In a fallback action, we observe that the command has the two following aspects in their command field:
- The first command contains an error related to RE.
```
[buck2/app/buck2_event_observer/src/last_command_execution_kind.rs:33] first_command_status = Some(
        Error(
            Error {
                stage: "remote_upload_error",
                error: "Remote Execution Error (reSessionID-763a0f96-98dd-4cc6-bb8a-dfeef09f4a2b): Error during call: Code: 3, Message: The given use case id=something-bad is not found.",
            },
        ),
    )
```
- The last command contains an execution kind of "local".

```
[buck2/app/buck2_event_observer/src/last_command_execution_kind.rs:41] last_command_kind = Some(
    OmittedLocalCommand(
        OmittedLocalCommand {
            action_digest: "4315fb63178629c5a8b80e8e7906743aeb2e67ab:94",
        },
    ),
)
```

If these two conditions hold true, then we infer that the action was a fallback action.

Reviewed By: krallin

Differential Revision: D42749228

fbshipit-source-id: 00e57170cf456b7260ba3102a8dd5aa4be36212e