Add fallback field in summary count when actions use fallback
Summary:
This diff continues the efforts outlined in D42749228, which is to add UX improvements to buck2 when an action falls back from RE to local.

Specifically, this diff introduces the `Fallback: {}/{}.` to the summary count when a job is running.  Note that this field only appears if there is a fallback action among the local actions.  If there are none, this field will not appear to avoid clutter.

Reviewed By: krallin

Differential Revision: D43011755

fbshipit-source-id: ef075aa5baf5eef771a1a85fa48ff653b3af5452