Add action fallback count to buck2_builds table
Summary:
Now that we are surfacing the number of actions that use fallback on the buck2 user-facing CLI, let's log these values over time as well, as they may prove useful for detecting abnormal behavior when it comes to using RE.

This diff adds `run_fallback_count` to our buck2_builds Scuba table by adding the count in via Ingress.  It also moves the function `was_action_fallback` outside the ActionStats implementation, as it doesn't actually require a reference to `self` anyway.

Reviewed By: krallin

Differential Revision: D43197531

fbshipit-source-id: 9f538bb8db07655c3b68107d04c688af8fbd5927