Don't specify input.lib.dex as an input for pre-dex-merge dynamic action
Summary:
This was pretty wasteful - we don't actually need these for the dynamic action, and I was observing a pretty long gap in `buck2 build fb4a` while we waited for materialization so that the dynamic action could run.

A couple of things:
(a) Perhaps `buck2` should warn you if you specify artifacts that you don't then read?
(b) We should probably resurrect T128153110, since a long period without any console output makes it look like nothing is happening, which is bad. At least if it told you it was materializing artifacts (or whatever else) you'd know something was happening and also would have more information to try to improve things.

Reviewed By: navidqar

Differential Revision: D43271444

fbshipit-source-id: f6246fccad6bce000ec9ba7b0fb57b754cc8d8e9