buck2: ignore digest type in Buck -> RE conversions
Summary:
Like it says in the title. Our plan to support configurable (and potentially
mixed hashes) is to configure what hashes get *into the system*, so it follows
that the output should not care about what format it has and just faithfully
emit the bytes.

Does that suffice? We'll have to see! It does for supporting backends that only
support one digest type (in which case you've got to make sure all your inputs
produce the right hash format), but it might not for backends that support
mixed hashes (if those e.g. require you some disambiguation), but those a)
don't exist right now, and while they will exist when RE starts supporting
Blake3, we can b) influence that design to make this work for us.

Differential Revision: D43054891

fbshipit-source-id: 8044f2a867fc413fd21e20aa65a6552b96a05fa4