buck2: clarify a bit of code
Summary:
This code *looks* like it uses the directory digest for something that isn't
just debugging, but in reality it doesn't. Rewrite to clarify this is only
for debug + stop inserting the root directory in there if you're not gonna
use it.

The generics aren't just for show here, later I do use this.

Reviewed By: stepancheg

Differential Revision: D43054884

fbshipit-source-id: 756ad861522cc3a49665a69ea33c183a75b4ae26