buck2: cas_digest: rename some structs
Summary:
`DigestHash` is technically sort of correct, but it's weird to have `CasDigest`
have a method called `digest` that gives you a thing that claims to be a "hash"
(but on the flip side naming a method `hash` leads the Rust compiler to give
very questionable advice if you typo it). So, let's make it `RawDigest`.

`DigestKind` is also kind of meh considering we also use "kind" to mean file
vs. action cache entry, so let's just call it "algorithm".

Reviewed By: stepancheg

Differential Revision: D43054886

fbshipit-source-id: 57ceb6c67a85b3c0630069bd5db051650e10f3af