buck2: thread a "DigestConfig" to all RE -> Buck2 digest conversions
Summary:
This doesn't currently do anything, but the intention is to use this to inform
how to interpret digests we get from RE. For example, this can be used to tell
whether a given hash is SHA256 or Blake3.

In an ideal world, this would be part of the GRPC protocol, but the Bazel RE
spec didn't have this and so OSS implementations don't either.

Differential Revision: D43054893

fbshipit-source-id: 2668545c84080e65731f9376c630168e47e19e21