Enable migrating to correct cell resolution for query
Summary:
Currently, bxl queries use the wrong cwd for cell resolution. It's using the current working directory that the user is in, when it should be using the current working directory of the BXL script.

I'm adding a flag to be passed into cquery/uquery so we can do a migration.

The plan:

* Land this and the next diff
* Migrate all use cases to use `cquery_legacy()`/`uquery_legacy()`
* Wait a day or two for things to stabilize
* Fix `cquery()`/`uquery()` impl
* Wait a day or two for things to stabilize. At this point, both the correct and incorrect implementations are in the code base
* Replace all use cases to use `cquery()`/`uquery()` again
* Remove `cquery_legacy()`/`uquery_legacy()` from code base

Reviewed By: bobyangyf

Differential Revision: D43243352

fbshipit-source-id: 228b8f457512097e3d823e32be9d0d8486b0d7b8