Set __file__ attribute on extension modules
Summary:
Set the `__file__` attribute by default to `static-extension`
We could set it to the `executable_name` but that is incorrect for extensions in link groups.
I played around with using `dladdr()` to look up the library name, but that is more complicated than we need for now. [dladdr](https://man7.org/linux/man-pages/man3/dladdr.3.html)

Reviewed By: andrewjcg

Differential Revision: D43173279

fbshipit-source-id: bdfb9fcd01f43ab44e40975c6b5fd620680f5c90