KotlinCDMain can produce a class ABI's dir representation
Summary: KotlinCDMain first runs `StubJar` to create a class ABI `.jar` from the library output. After that, it will iterate through the `.jar` and for each entry it will create a file in the ABI output dir at the same path. Rather than extracting the whole file, it just writes the hash of the file (since its cheaper to do that).

Reviewed By: christolliday

Differential Revision: D43158272

fbshipit-source-id: 4f86a5c7ea13e1bd26d0263b78ca028eaf0a8e4d