buck2: http download: don't assume you're hashing into SHA1
Summary:
Like it says in the title. This used to assume the FileDigest was going to be
SHA1, but that's no longer the case, so let's rework this. Instead of
explicitly targeting SHA1 and SHA256, we create a validator for each checksum
and then either look at the FileDigest if it's the right kind, or create a new
hasher.

Reviewed By: bobyangyf

Differential Revision: D43279726

fbshipit-source-id: b58f6cdd3f1f289694af1526c3e1c4b9c782431e