Fix force-static heuristic (take 2)
Summary:
Libraries will mark themselves as `preferred_linkage = "static"` to
ensure they always get linked into their dependents, which omnibus
honors in it's implementation: if a root node depends on a force-static
body node, that body node is statically linked into the root node; if
no other body node depends on that force-static node, then it will not
be added to the link line of the body.

Respect this heuristic in link groups, but always statically linking a force-
static node into it's dependent, even if it crosses a link group.

Reviewed By: christycylee

Differential Revision: D43278287

fbshipit-source-id: d59b7303bd5c26cc6b041db02d4ec4c725ec38e6