buck2: unhashed_outputs: handle projected artifacts properly
Summary:
Right now we ignore the projection in the unhashed path, but not the output
path, so predictably that is not good because that means we symlink to projected
artifact but do it using the not-projected artifact's path.

So, if you have `out` projected as `out/foo`, we create a symlink called `out` and
point it to `out/foo`.

Let's fix it by ignoring it in both, meaning we'll create a symlink called `out`
and point it to `out`.

Reviewed By: bobyangyf

Differential Revision: D43368010

fbshipit-source-id: 1368f198e31e54ee63f77fa1a13b4e539c8199b2