e2e: fix test_js test
Summary:
Some of the targets are not building due to the output type not being correct. (we expect directory but it is not)

Introduced in D43328609 (https://github.com/facebookincubator/buck2/commit/f6804a9f3a35fd87ee34476fdb6f25dd9b0997ce), but since some of the dir=true markings are correct, fixing it forward.

Reviewed By: blackm00n

Differential Revision: D43398283

fbshipit-source-id: ec89c7dcfc4be586990f3faf9e2606a3fa166ce1