buck2: OSS: executor config: restructure it
Summary:
Currently, there is no way to represent how caching gets enabled or not in
your executor config, which works fine internally (we always set it even for
platforms that can't possibly cache things), but does not work at all in
open-source (where we really need to not make it the default).

There is also a recurring issue we have run into with the way we configure RE
use case and RE properties for the caching executor, which is that we pull them
out of the underlying RE executor. That is very messy and often times leads to
surprising behavior when flipping `remote_enabled = False` changes all your
cache keys.

To fix both of those issues, let's restructure our executor config to look
more like what it actually needs to express, which is that your executor is
either:

- Local only and does not interact with RE at all
- Remote enabled, in which case it might interact with RE to either / or:
  - Query caches
  - Run actions

This is what this diff does. Immediate effects of this are:

- In OSS, you get a truly local executor by default
- Internally, for e.g. BXL, we stop querying caches that are guaranteed to not
  have the data you want them to have.

Do note that currently in OSS, we disable caching *always* so you already were
getting a "truly local" executor, but that is something we need to remove
because we do actually want to enable using RE caching in OSS, which right now
is not possible!

Reviewed By: ndmitchell

Differential Revision: D43349777

fbshipit-source-id: 30c21735208f697c6e02554d6c37e87f5a743e75