buck2: make re_platform a property of CommandExecutor, not PreparedCommandExecutor
Summary:
That is where it _actually is used_, so move it over where it makes sense.

Also, because it is important all layers of the executor stack agree on what
the re_platform is and the best way to guarantee that is to only have it exist
in one place.

Reviewed By: ndmitchell

Differential Revision: D43349781

fbshipit-source-id: 4f2eb4efe3b0487d5ac77e27875b8eee8ef500f4