buck2: simplify getting default executor config
Summary:
There is basically zero value in this nowadays because setting `--local-only`
actually works with a hybrid executor, so all this trying to define a default
executor that would work is just not very useful at all.

We lose out a little bit on warning the user that `--remote-only` in OSS makes
no sense if you didn't configure execution platforms (because we cannot come up
with good defaults for you there), but that's probably fine since we error
out fairly clearly later.

Reviewed By: ndmitchell

Differential Revision: D43349782

fbshipit-source-id: ed73cdd37542a8ada4efba57d49bda798729fb6b