buck2: OSS: remove vendored OpenSSL
Summary:
We don't need this anymore as of the previous diff, and having 2 TLS
implementations isn't great.

Whether Rustls or OpenSSL is "better" is debatable, but we *must* use Rustls
currently since that's the only option for Tonic, so we might as well reduce
our overall maintenance burden and just standardize on Rustls.

Reviewed By: ndmitchell, stepancheg

Differential Revision: D43434495

fbshipit-source-id: ed7fd0b79fb0b8a96a61202545d7e756f389cf1e