buck2: make output paths conflicts a hard error + make it easier to debug
Summary:
Like it says in the title. Hard error when this happens, because we don't
have any more instances of this in the repo (but we do have an instance on a
diff from native Python).

Also make it easier to debug because that is a nice thing to do when you error
out by telling .

Differential Revision: D43458591

fbshipit-source-id: 8ca983afc330ae0dd8f14e8431565b5056bf607e