buck2: materializer: deferred: make the version API more robust (reland)
Summary:
Reland D43000548 (https://github.com/facebookincubator/buck2/commit/d20d05bcf5d696841f887cdc447f1e1d0988e62c). This was the diff that had a bug. We inadvertently skipped
updating the status of the processing field when materialization finished on an
already-materialized entry ( = a deps-only materialization). This updated
version corrects that bug and a test is on the bottom diff.

Like it says in the title. Versions are used so that when a processing
future calls back, we can tell whether it was current, so it follows that
we always need to update the version when setting the processing future, so
let's bake that into the API.

Reviewed By: IanChilds

Differential Revision: D43469789

fbshipit-source-id: a469e22985f024d30a31ef14a879d2f48ba3d99a