relax the no transitive-deps rule
Summary:
Before D38213812 (https://github.com/facebookincubator/buck2/commit/d872b98bd4537eb99262fcfca035fd148e515137) if a target *t* depended on a library *a* where *a* depends on a library *b*, then the modules of *b* were accessible to *t*. In D38213812 (https://github.com/facebookincubator/buck2/commit/d872b98bd4537eb99262fcfca035fd148e515137) we changed this behavior so that in order to use the modules of *b* from *t*, then *b* must be in *t*'s `deps` (that is, a "direct" dependency).

Why? williewillus [noted](https://fb.workplace.com/groups/403065016408786/permalink/8781197225262148/) buck2 differed from buck1 in this respect, admitting transitive use where buck1 did not. I argued then that we wrote the buck2 build rules that way with the understanding we were overcoming a buck1 shortcoming. However opinion prevailed that the buck1 behavior was "better". For example, mroch argued "I was able to depend on `String_utils` in `Server_files_js`, because `Server_files_js` depends on `Sys_utils` and `Sys_utils` depends on `String_utils`. That's an implementation detail! I should have to explicitly depend on things I use directly, not get them for free because something else used them transitively".

We found cause to reconsider this restriction.

Consider this program:
```
  let (): unit = Core.Printf.printf "Hello world!"
```

The dune file for the above program would look something like this:
```
(executable
 (name core-test)
 (libraries core)
)
```

When we translate that into a buck2 target we get the following.
```
  ocaml_binary(
      name = "core-test",
      srcs = ["core_test.ml"],
      external_deps = [
          ("supercaml", None, "core"),
      ],
  )
```

Before this diff (in the presence of the no transitive deps rule) building this target results in an error:
```
  1 | let (): unit = Core.Printf.printf "Hello world!"
                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  Error: This expression has type Core__.Import.unit = Base.Unit.t
         but an expression was expected of type unit
```

The issue is that for this to work there needs to be a `-I` for `Base` on the compile line and we deliberately prune it out! So, this is a fix
```
  ocaml_binary(
      name = "core-test",
      srcs = ["core_test.ml"],
      external_deps = [
          ("supercaml", None, "core"),
          ("supercaml", None, "base"),
      ],
  )
```
but it's not satisfying and potentially hard to discover. If on the other hand we relax the no transitive deps rule, the build "just works" as it does with dune.

Given this example, as samwgoldman put it, adding `-I`'s for transitive dependencies is the "lesser of two evils".

*Note: There are ongoing discussions in the community about compiler support for solving the problems we are grappling with here. See e.g. https://github.com/ocaml/dune/pull/430#issuecomment-423355545 and https://github.com/ocaml/RFCs/pull/31.*

Reviewed By: samwgoldman

Differential Revision: D43504245

fbshipit-source-id: ccb70c4b6c8e700287e1c5eb8eff0677a8b7f855