add buck2 library support for julia
Summary:
Here, we add two library rules for Julia: (1) standard libraries, where all the code is native Julia code; (2) `jll` libraries, which interface with c/c++ artifacts.

To accomplish this, we leverage `transitive_set`s to build up a tree of all the dependencies. For each dependency we build a sym_directory, and store that directory in the tree. Then, we add all of these sym directories as hidden dependencies to our command line *and* we update our JULIA_LOAD_PATH to point to this sym directory location.

For the `jll` rule, we dynamically build a custom package in ~~`buck-out`~~ a temporary directory that contains all the proper artifact mappings. This library can then be imported like a standard `jll` package.

To accomplish all of this, we use a python bootstrap rule.

(krallin, thanks for all the help you've provided!)

Reviewed By: krallin

Differential Revision: D43211037

fbshipit-source-id: 9d73870f52e8e5d83e6ee91f831251f6aa18a817