audit output fix for thrift_library
Summary:
Previously, audit output was constructing the project relative path out of an action's outputs, and checking that the inputted buck-out path's section after the isolation directory was the suffix of the project relative path.

For example, if an action's output's project relative path is

```
buck-out/v2/gen/fbcode/cfg_hash/foo/bar/__target__/out/blah.txt
```

and the inputted buck-out path is:

```
buck-out/isolation_dir/gen/fbcode/cfg_hash/foo/bar/__target__/out/blah.txt
```

we would check that the action output's relative path ends with:

```
gen/fbcode/cfg_hash/foo/bar/__target__/out/blah.txt`
```

This is fine, but sometimes rules will declare the output as a directory. So in this example, the inputted buck-out path specifies a file within the output directory.

So, if an action's output's project relative path is:

```
buck-out/v2/gen/fbcode/cfg_hash/foo/bar/__target__/out
# Note that the output is declared as a directory
```
then audit output will not return anything because we are trying to do a full suffix match.

Let's change this behavior to just checking that the declared output of the action is the prefix of the buck-out's artifact path. Using our previous example, we would be checking if `out/blah.txt` starts with `out`, which is the output directory declared by the action. Note that I'm using `starts_with` provided by the `ForwardRelativePath` as opposed to `String::starts_with` so that we make sure we are not accidentally breaking up the directory names.

Differential Revision: D43544051

fbshipit-source-id: 961c23d5c2f6d096bc4464ff1b098b6bcdfd21d3