chore: add boilerplate for hacking on buck2 with nix + direnv
While hacking around the buck2 codebase and testing builds, I needed
the tools necessary to build it. Most people use rustup, but Nix users
like myself tend to use, well, Nix, and don't keep cargo or other tools
always installed in PATH.

A popular way of doing things in the Nix world is instead to describe
the tools needed to build your package, and then let Nix put them into
your `$PATH` temporarily in a subshell. This allows you to have turnkey
development environments by simply running 'nix shell'. We do exactly
this, using a nix `flake.nix` file to describe what buck2 needs to build
-- mostly, just Cargo and the nightly version of rust specified from
`rust-toolchain`.

On top of that, direnv is a popular tool for automating this task, and
can be integrated with Nix. So we add two and two together.

Assuming you have installed Nix and direnv, you can simply do the
following to build buck2:

    $ cd path/to/buck2
    $ cargo build --release --bin=buck2

You do not need any previous Rust toolchain for this to work. There
could also be more tools provided as well, without the need for a
developer to ever install them -- such as cargo-bloat or cargo-fuzz,
as an example.

Note that the particular rustc that is put into `$PATH` above will
be read directly from the `rust-toolchain` file. We use a Nix overlay
called 'rust-overlay' which provides nightly compiler builds derived
from the official upstream binary releases of the compiler. This overlay
is updated daily with info from nightly builds. Sometimes it may be
necessary to run 'nix flake update' before the necessary nightly version
can be used.

A follow up will actually create a buck2 expression written in Nix,
which can be used by anyone to build a copy of buck2

Signed-off-by: Austin Seipp <aseipp@pobox.com>