cli: Remove `#![feature(box_syntax)]`
Summary:
This feature is [going away upstream](https://github.com/rust-lang/rust/pull/108471). Let's remove it from buck2 before we cause Andres pain when he upgrades the Rust toolchain.

I'm replacing the uses of `box` here with just `Box::new`. `box` is technically supposed to have an additional optimization, on top of the function call (specifically in-place construction), but I am not worried about that for a couple reasons:

 1. Most of the uses of this in buck clearly didn't need it.
 2. This [did not always do what you expected it to](https://play.rust-lang.org/?version=nightly&mode=debug&edition=2021&gist=5ba659f703a301ae50d87e1c04d417aa). The optimization was pretty specifically designed for `box [0; 1_000_000]` and that's more or less the only thing that ever worked.
 3. The alternative that allows us to retain this optimization is an unstable `#[rustc_box]` attribute. However, this would require turning on `#![feature(rustc_attrs)]` - even as unstable features go, that one is super dangerous and super unstable. We really should not use it.

If there is a case where this really matters, people are free to use `Vec::with_capacity`, the allocator API, etc. to still get in-place construction.

Reviewed By: bobyangyf

Differential Revision: D43683120

fbshipit-source-id: cffbd50f275f8cc43591aac70d413363dabcd0de