buck2: InvocationRecorder: inline matches
Summary:
We basically implement an "unpacking event recorder" that is 90% what the
InvocationRecorder (and only it) need. Let's just have the logic for the
InvocationRecorder live there.

This means less boilerplate when adding things the InvocationRecorder cares
about, since we don't need to declare the methods twice and implement the
match in a completely different module.

Reviewed By: stepancheg

Differential Revision: D43697915

fbshipit-source-id: 83c395f98e23578727640cfd463db7d2bdd83dfd