buck2: subscribe: exit when the client is gone
Summary:
Like it says in the title. When stdin is gone, gracefully disconnect.

We don't really *rely* on this being graceful server side (StreamingRequest
exits when the stream ends, we could just hang up), but it doesn't hurt to
have, but this is the easiest way to terminate everything gracefully and
get our CommandResult back.

Reviewed By: bobyangyf

Differential Revision: D43667769

fbshipit-source-id: 990a1a7422e42a230bf05b54b352ff698ae35865