buck2: don't reuse a daemon if the Sandcastle ID is different
Summary:
On Mac CI, we see jobs reusing Buck2 daemons created by a previous Sandcastle
job. There is no reason for us to reuse a daemon with a difference Sandcastle
ID, so I'd like to propose rejecting such a daemon (and creating a new one).

The way we do this is we hook into our existing version check, and check
the sandcastle id in addition to the version.

This does mean Sandcastle needs to actually set this whenever it calls Buck2
but that's something it's pretty good at overall.

Differential Revision: D39472624

fbshipit-source-id: 114752208315b70476925a1299eefc6bac473554