buck2: json logs: show durations as micros
Summary:
Right we render durations in JSON logs as `[seconds, nanos]`, which is really
annoying to work with (because you can't just easily sum those).

Let's just standardize on using microseconds, which is what we use for CPU
durations already. Since we no longer rely on format stability from JSON logs,
this is easy to do.

If someone was depending on the exact output of `duration` in our
JSON-formatted logs, they'd be broken by this, but I couldn't find any
instances of this (though obviously I can't look at people's scripts).

Differential Revision: D43770032

fbshipit-source-id: 4efa4432968f899f2b796712470ebc0300ab4cd4