buck2: don't retry output missing errors locally (and fix a broken test)
Summary:
This test is failing because RE now throws on missing outputs, and that leads
us to retry those locally (we should fix that and make this error not
retryable, but for now let's fix the test).

Let's update to avoid retrying locally. The way we get this output is a little
awkward. In an ideal world, RE would just not cache their result and would
return normally to us, but we're not in that world so for now just handle
this the best we can.

Reviewed By: lmvasquezg

Differential Revision: D43833295

fbshipit-source-id: f05d3a3e3c35801f3e57dbabebab30f6fb53f5c5