Fill in a stub for wait_for_idle
Summary:
This isn't really what we are supposed to do.
But it suffices for non-racey command uses cases of buck2 for dice to behave like this.
this is used to make sure we wait until all previous tasks have finished before executing the next command. This would make us fail to wait, but usually, that's okay.

Reviewed By: stepancheg

Differential Revision: D43824891

fbshipit-source-id: 4824dcc95283ad4c0ceb15d74165aa2cb631a522