buck2: refactor CommandExecutionTarget
Summary:
We use this in both tests and action execution, and as a result we end up
having to have tests pretend they are actions, which causes a bunch of issues
around tests having to pretend they have a category (they don't) and a unique
identifier (they *definitely* don't), which they don't actually need.

This fixes this by making the command execution target basically a trait
(modulo some cleanup that I'm doing later in this stack) and letting actions
and tests provide a different implementation.

Differential Revision: D43905187

fbshipit-source-id: d21287d98c88ddd6c72c9a43137a89c09105a707