classpath_jar_tag should only tag the compile classpath
Summary:
Previously, we weren't able to call `tag_artifacts` on a JSON projection, so we had to tag the whole `proto.json` and then always tell `buck2` that all the non-classpath inputs to that file were always used. This was buggy because we missed some inputs.

Instead, we can now just call `tag_artifacts` on the `classpath` itself, and then use the output from the compiler to say which of the `.jar`s on the classpath were actually used. This is much simpler.

There's a bit of refactoring required because the place where we run the command and the place where we set up the classpath are different, but it's still pretty simple, we just have to pass around the `classpath_jars_tag` to the relevant places.

Reviewed By: krallin

Differential Revision: D43871801

fbshipit-source-id: c93bc857fce35b082de64bf654f76630358e4e81