Add "headers" subtarget for apple_resource
Summary:
# Context

I’m working on using buck2 to build fbios dylib build. With dylib build, one major difference is that we rewrite some of the target’s dependencies. For example, if FBFoo depend on FBBar, and FBBar lives in a different dylib, FBBar will be rewritten as ":FBBar#headers" during build.

# This diff

During the rewrite, I only have access to the dependee's attributes, not the depender. So I cannot really skip rewriting the `apple_resource` target.
So the easiest fix here is adding a dummy `headers` subtarget that produces no output.

Reviewed By: IanChilds

Differential Revision: D43911967

fbshipit-source-id: e3dc402eedde911b3c479cfc8bbef76df2059cf1