Add param to allow overriding/transitioning configuration
Summary:
Support a `constraint_overrides` parameter for the buck2 Python executable
rules, which callers can use to add constraint that should be used to override
the rules configuration (forcing a transition).

In practice, Python rules can use this to allow per-target control over the Python
version that it's built with, which may need to be propagated to dependencies
(e.g. native extensions need to build against a matching libpython).

Reviewed By: scottcao

Differential Revision: D43932867

fbshipit-source-id: 0c42d162e63f33cdf35135784019e80aa3a42024