buck2: deferred materializer: refresh individual digests
Summary:
We were originally recommended to refresh by querying the action cache,
but a problem with this is that it may not work when an action is
non-deterministic and has been re-run despite not having expired (which is
rare but happens somewhat frequently at scale).

Namely, this happens:

- You run A on RE, get X as output
- Someone else runs A on RE, gets Y as output
- You fresh A, you now refresh Y (but you think you're refreshing X)
- X expires and you get an error

This diff fixes that by freshing individual blobs and not actions.

Reviewed By: christolliday

Differential Revision: D43916180

fbshipit-source-id: 01a6af2ec41633b4bdcb7a1c3d5e73ef7b3697a2