buck2: stop using `when XXX` in `context()`
Summary:
This would make sense if we displayed our errors like so:

```
format!("Error: {:#}", err)
```

Which would look like:

```
Error when doing XXX
```

However, we don't, so our output ends up looking like this:

```
when XXX

Caused by:
  Foo
  Bar
```

This is kinda confusing. What does `when XXX` mean?

E.g.:

```
[torozco@devbig053]~/fbsource2 % buck2 build blah
when gathering package listing for `fbsource//blah`

Caused by:
    0: The package is invalid
    1: This error was caused by the end user
    2: Error listing dir `fbsource//blah`
    3: Eden POSIX error: blah: No such file or directory
```

Let's just not do this and actually inicate this is an error.

Differential Revision: D44020500

fbshipit-source-id: b35b7e42c16241078c243d0a167d5edf0347109e