buck2: make buck2.sh support `buck2 test`
Summary:
To support fast rollbacks the repo specifies the path to Tpx as relative to
the Buck2 binary. However this means when you just build your own Buck2, it
will not find Tpx.

We can support this in buck2.sh by creating a dir with the same layout as the
one we put in the release binary. One problem is that we'd have to copy the
binaries to do this which is adds a little overhead.

To avoid this, add support in buck2 to receive an override for this path.
On platforms where this might be harder to do (Windows) we can
always just use copies if we want (our binaries are also smaller there).

Reviewed By: shonaganuma

Differential Revision: D44003531

fbshipit-source-id: 8b4d7167ceaa59eac083d960a371f6f76857df7d