buck2: invocation recorder: send the command name
Summary:
Because we have it and because this way we don't have to infer it from
CommandEnd, which is a) verbose and b) might not actually work if the
CommandEnd event wasn't ever sent, which will happen on cancellation.

We could switch to CommandStart but a) still applies and b) still kicks in in
some cases (early crashes, cancellation while waiting for DICE).

Reviewed By: IanChilds

Differential Revision: D44130093

fbshipit-source-id: cbf0288e68c85f265cfb1106ede2846ec2eb0b7a