buck2: move `trace_id` to `ClientContext`
Summary:
It actually makes no sense to pretend that we create a new one per command
context, because we always get this from the wrapper. Move it up the stack
and then the rest of the code can stop pretending that we don't know the
TraceId for our command until we get an event.

Reviewed By: get9

Differential Revision: D44134953

fbshipit-source-id: 2b28ad89edb0c9795a4c1e58250a282d58eaa99a