buck2: forkserver: wait() `Child` before dropping it
Summary:
So that on cancellation we don't leak anything. In practice on Linux I don't
see us creating zombies when we cancel, so I don't think there *is* an issue,
but it's worth doing the right here (the docs point to this being the right
thing):

https://doc.rust-lang.org/std/process/struct.Child.html#warning

Also makes debugging easier, since now the next diff hangs in a more
logical place rather that somewhere in Cargo.

Reviewed By: lmvasquezg

Differential Revision: D44165674

fbshipit-source-id: 9f69549a71c56c18826e8235dc80d7d767a76f32