Obtain coverage flags from buckconfig
Summary:
This diff is partially based on zertosh's prescient comment in https://www.internalfb.com/diff/D41603520?transaction_fbid=894031748435274.

> "-Cinstrument-coverage" -- Why is this not in a mode file?

> Just do this and I'll clean it up later.

Tying the flag to a mode file allows differentiating between LLVM 12-based C++ builds (for which we support coverage, because rustc is currently also being built against LLVM 12) and LLVM 15-based (for which we do not).

Fixes T147052143: errors on platform010-clang-15 when using clang built for LLVM 15 and rustc built for LLVM 12, and coverage being enabled on both.

```
7886794136458710398_3.cdata: malformed instrumentation profile data: counter offset 200362944 is greater than the maximum counter offset 374959
error: no profile can be merged
```

In the future if we build a separate rustc for platform010-clang and platform010-clang-15 (as is done for clang), we can re-enable coverage on the second one.

Reviewed By: zertosh

Differential Revision: D44155933

fbshipit-source-id: 97f2576826ed1f7839b8787774006021189db34a