Replace BUCK2_ENABLE_SCRIBE with BUCK2_TEST_DISABLE_LOG_UPLOAD
Summary:
Before this stack of diffs lands, it's possible to control scribe log uploading and event/re log uploading separately by 2 different env vars. This doesn't make sense, rather, increases the complexity because when we are in the situation where we don't need the logs, we don't want both.

By this stack of diff, we will be able to switch on/off everything using `BUCK2_TEST_DISABLE_LOG_UPLOAD` flag. In turn, `BUCK2_ENABLE_SCRIBE` will be deprecated.

Reviewed By: krallin

Differential Revision: D43999771

fbshipit-source-id: 716564e8b266d2dbe5b8bb0894d2f6604c6f0966