buck2: invocation recorder: add metadata in ... one more place
Summary:
We currently haves copies in this in way too many places, namely:

- CommandStart
- CommandEnd
- ComandCriticalStart
- CommandCriticalEnd

A variety of those places have access to different metadata. It's a bit messy.
What I am going to do here however is simplify so that we merge it all in one
place and send that to Ingress.

For now, add scaffolding.

Reviewed By: get9

Differential Revision: D44173706

fbshipit-source-id: db5ee0f0ab3a17b24a924f5a893b2c9ced57b473