Instantiate InvacationRecorder even when Scribe is turned off
Summary:
Currently, `InvocationRecorder` is not instantiated when Scribe sink is disabled e.g. when `BUCK2_ENABLE_SCRIBE=0`. This causes a problem because `InvocationRecorder` provides some other feature interfaces, for example an error categorizer that can be seen on next stack diff, than just sending a record event.

By this diff, buck2 client always instantiates `InvocationRecorder`. When Scribe sink is disabled, only the difference in its behavior is that it doesn't send out the record event at the very end of its life.

Reviewed By: stepancheg

Differential Revision: D43019530

fbshipit-source-id: ba40d8f12834fcb6a9a040cd0b788c197200edc1