Make exit code in TestResponse optional to distinguish build error and test error
Summary:
Currently, if `buck2 test xxx` didn't execute any tests because e.g. the test binary compile failed, buck2 daemon always sends back "1" to the client and the client obeys it, always returning 1 as an exit code. It's true that the client should respect the test executor's exit code that the daemon reported, but if it's not the case, the client should control over the exit code determination.

This diff changes the `exit_code` field in `TestResponse` into `optional int32` so it can express "Test executor didn't yield an exit code. Delegating you to generate one." by setting `None`.

In this diff, the client calls `ExitResult::failure()` when `None` is returned back. In next upcoming diff where the categorizer is implemented, the `failure()` will trigger the error categorizer and the exit code will get updated later on.

Reviewed By: krallin

Differential Revision: D44178196

fbshipit-source-id: d20fd0c00708a6466068b69b3cf5941292a20966