offline builds: Re-organize daemon startup configs originating from command line options
Summary:
So we can better capture mode files used in offline archives, we need to be able to start a daemon up with I/O tracing (because mode files are read outside the I/O recording layer).

This re-organizes some code to make it easier to plumb through the "enable I/O tracing" option on daemon startup.

Reviewed By: krallin

Differential Revision: D44208130

fbshipit-source-id: c56528d0caa7c8bd1b986c7658f46364b7272e8b