offline builds: Skeleton for intracting with i/o tracing
Summary:
This diff implements the skeleton for I/O tracing of the daemon; this will eventually power having buck2 provide an offline archive (or a manifest such that we _can_ produce an offline archive) to be able to do builds in a fully offline/no-RE environment.

This implements entrypoints for enabling and disabling I/O tracing, as well as clearing the internal tracing state (e.g. resetting the internal buffer and emptying it).

Hypothetical way to get buck2 to provide a manifest for an offline archive:
```
$ buck2 debug io-tracing enable
...
$ buck2 build //my:target
...
$ buck2 debug export-offline-archive /path/to/an/archive.json
...
$ buck2 debug io-tracing disable # or clear
...
```
Then some other non-buck2 binary could go in and assemble the archive into a zstd-compressed blob.

Reviewed By: krallin

Differential Revision: D43512525

fbshipit-source-id: 9ff303ba248c5b4ded7a80c4e43574afed0b0e40