offline builds: Export I/O trace as a manifest with some additional information
Summary:
Exports the I/O trace as a well-formatted structure. Currently we include:
* The list of paths accessed by builds
* The repo revision (still tbd)
* The path to the buck2 executable (currently hardcoded). In actuality this can probably stay hardcoded since we want to get the actual dotslash file

After this, we'll hand off the manifest to a separate binary that will transform various file paths (e.g. materialize dotslash files) and package this up into a single zstd compressed archive. This is introduced in later diffs.

Reviewed By: krallin

Differential Revision: D44154910

fbshipit-source-id: b124076583c1450516040df19144a4985b3ca017