Add EdenFS version to buck2_builds Scuba table
Summary:
This diff adds `eden_version` as a column to the buck2_builds Scuba table.  The hope is that knowing EdenFS version will serve as useful information to potentially identifying trends amongst buck2 builds.

Note that a version only logs if the IoProvider is Eden, and if the filesystem is detected to be an Eden one.  For all other cases (e.g. IoProvider is `fs`, filesystem is not Eden), it will contain a `null` value.

Reviewed By: krallin

Differential Revision: D39988267

fbshipit-source-id: be69a022d216e27589ee39b2cba370c535490fb1