buck2: add `killall`
Summary:
Like it says in the title. This uses the same implementation was the wrapper.
The reason to have it (in addition to the wrapper: see next diff) is to let us
log those calls.

This is also nice to have for OSS since we recommend this command in quite a
few error messages.

Reviewed By: stepancheg

Differential Revision: D44217163

fbshipit-source-id: 3c5155bd284d9e5e22ebc71fb2b2b4c38f0e20d2