Calculate active keys differently for modern dice
Summary:
This is technically not so accurate, but the accurate calculation is expensive and freezes up the shared-dice cache, which results in large regressions for modern DICE

By doing less precise metrics gathering, we improve perf from 10m => 2m on a fully cached, non-materialized fb4a build.
this is slightly higher than legacy dice at 1m45s, but spends much less memory.

Sacrificing some preciseness of metrics for now until we figure out a better-performant way to collect it seems a worthwhile trade-off

## why
So modern dice calculation for this is more complex than legacy because the per-transaction map is no longer of only currently active keys.
Additionally, previously we would hold a Dashmap for each Key type, while we now hold one single dashmap for all keys.
This introduces much higher cost of contention if we spend time locking the map for metrics, because instead of only specific key types at a time being locked, we now lock all keys

Reviewed By: krallin

Differential Revision: D44445716

fbshipit-source-id: e209218c320666d3669beafcd98a6d5369f27dec