Iterator -> IntoIterator
Summary:
Use `IntoIterator` in function parameters instead of `Iterator`.

Practically does not change much, but this is how these types are meant to be used: `Iterator` represents iteration state, while `IntoIterator` is a "collection" interface.

For example, when calling `get_compatible_targets` with `loaded_targets` parameter, `loaded_targets` is a collection, not middle of iteration.

Also occasionally, this makes code shorter. For example, it is possible to pass `&[]` to a function which accepts `IntoIterator`.

Reviewed By: bobyangyf

Differential Revision: D44553531

fbshipit-source-id: bbf08ad39a938de92d1aede9ed85e9deff9f22ea