Fix bug in ctx.output.print
Summary:
Let's simplify how we print things, even though we do some more string allocation. People typically don't print that much stuff out, so I don't think it matters that much.

This also fixes a bug with printing where a list of objects is printed on a newline with each object, which is incorrect.

Reviewed By: bobyangyf

Differential Revision: D43948042

fbshipit-source-id: 35ef587b99ada4fc0040771b148cb625a473cd7e