Avoid auto-select conversion when running cquery TDs
Summary:
Our TDs currently hard-code the target platform, which causes
issues when walking down nodes with incompatible selects.
We should probably add a `target_compatible_with` to these, but,
until we move to `uquery`, this will end up suppressing these
nodes, and transitive deps, from TD.

So, for now, opt-out of this auto-conversion when our cquery TD
indicator is set.

Reviewed By: xw285cornell

Differential Revision: D44601783

fbshipit-source-id: a3e9865bd2eeb02825e009d075d48784d5a5aa22