Fix the timed_list test tolerance for good
Summary:
Previously if the test took > 50ms it failed, which often broke on Windows. Recent changes made it 100ms, which meant it broke on Windows less, but I still encountered a couple of instances, especially on CircleCI for open source.

This change means we use the time speed, so we multiply all durations by 1000, and run the speed at 0.001, so the real elapsed time (which we'd like to be approx 0) gets divided by 1000. That means we have gone from 100ms of tolerance to 50s of tolerance.

Reviewed By: bobyangyf

Differential Revision: D44597625

fbshipit-source-id: 44a4795149ffe63129a3d24c53bd3ce14665616d