buck2: build_listener: send it total time
Summary:
Like it says in the title. Let's actually also report the total time here,
because ultimately that *is* how long the build took. This is particularly
important on builds that get a lot of cache hits, where right now we'll report
something like "ah yeah writing this file took 150us and *that* was the
critical path" whereas the build as a whole took a couple minutes.

We'll still report both times separately, but the end goal is to be able to
differentiate overhead from runtime here.

I didn't find a great name for "the time that isn't the total time", so I went
with "user".

For now, I'm not using it, but later in this stack I will change that.

Differential Revision: D44510934

fbshipit-source-id: 443e197c083820fca5e8c5433b7cabb688c0b177