buck2: build_listener: propagate "visibility" of actions behind top level analysis
Summary:
Consider this example:

I build a binary A that depends on library B where I need to compile a file C.

There's a single path here, so that's the critical path. Naïvely, it is:

- Build binary A
- ... depends on compile C
- ... depends on analyze B

But in reality, you didn't know you needed C until you finished the analysis of
A. So the real critical path is:

- Build binary A
- ... depends on compile C
- ... depends on analyze A
- ... depends on analyze B

This is what this diff does. We notify the build listener of what analysis
makes us request a given top level artifacts. Then, we inject some new edges in
our graph: anything those top level artifacts depends on gets an edge to the
first analysis that caused that artifact to become visible.

This is an approximation. It's possible that when you have multiple top level
targets, the first analysis that finishes isn't the one that ends up requesting
a given dependency first, but at worst it gives you an idealized view of how
short the critical path could be.

Reviewed By: bobyangyf

Differential Revision: D44510931

fbshipit-source-id: 83a5b2a9fc19b7914b2ff697b464166e654cc9e6