Fix up some of the py directory
Summary:
This used to fail with an error that it couldn't find subscription.proto. I've fixed that, but now it complains that protoc-gen-mypy needs upgrading to support optional types.

This definitely hasn't worked in a while, but this patch makes it more working. We should probably decide to kill or revive it though.

Differential Revision: D44641263

fbshipit-source-id: e14f254aa4f2d972cdbcbb5688e9b46388ac30cf