chrome trace: Count values in timeseries data per sec rather than per ms
Summary:
chrome trace time series data is measured in "count per ms" instead of the more intuitive "count per s".

This dimensionality change was made in D35795442. At the time, snapshots were emitted with the "burst" missed interval behavior, which resulted in, well, a burst of snapshot events if the timer was delayed. This then led to division by zero and ultimately caused chrome trace generation to fail.

The underlying behavior was fixed in D35821366.

Let's go back to measuring events per second instead of ms so values like bytes are a more reasonable value (instead of forcing people to scale by 1000x).

Differential Revision: D44642277

fbshipit-source-id: b7c94cb3c7300cbf57b5f453509e79e1dbb03202