Add attribute to discard link groups
Summary:
When using auto link groups and marking a link group with the
attribute `"discard_group"`, the link group will not be build at all,
effectively dropping it from the build graph.

This is particularly useful for dependencies that we statically know
cannot be reached, but cannot be removed from `TARGETS` file
due to other libraries depending on it.

Reviewed By: luciang

Differential Revision: D44562565

fbshipit-source-id: 198a5aa034db8ac681f1427fd91e5cea6c8e82d3