buck2: executors: re: support explicit cancellation
Summary:
Like it says in the title. When running a build, this just gets implicitly
cancelled when we drop the DICE task this is running on, but when running tests
we don't have that and we rely on an `LivelinessObserver` instead (which is
arguably a cleaner way to do this since it avoids surprise drops).

As a result, when running tests, we don't actually cancel things that are
running on RE, and instead we wait for them to finish. This means if you CTRL+C
a run that was running tests on RE, then run again, we'll have to wait for
your tests to finish executing instead of abandoning them immediately.

This diff fixes that by listening on our cancellation observer.

It would probably be a good idea to unify the build & test codepaths by
actually having `exec_cmd` enter `with_structured_cancellation` and then rely
exclusively on `LivelinessObserver` inside the executor, but that's a bigger
refactor so I'm not doing it just yet.

Reviewed By: bobyangyf

Differential Revision: D44667901

fbshipit-source-id: a54033fd6211404f7f6ac1d572cf40c6b316dc77