buck2: dep files: extract a function
Summary:
Both the "has match" and "has no match" branches need to take some follow up
action (on miss: clean dep files from the map, on match: validate with the
materializer) before they return, and an early-return style makes this very
easy to get wrong when refactoring.

So, split up the "does it match" and the "what to do if it matches" parts.

Reviewed By: stepancheg

Differential Revision: D44712620

fbshipit-source-id: e4967cbc66a6a8658ed2d1773519f00670d28034