buck2: actions: run: refactor some code
Summary:
The end goal here is:

- Use a single visit instead of one for dep files and then one for actual
  execution.
- Produce the inputs directory earlier in the codepath where we have dep files.

So that's what this refactor does. We give `prepare` a visitor and then it
does all the work and all our visitor has to do is be able to list the inputs.

I'm was only after the latter here, but since the `metadata_content` has to be
between listing inputs and creating the input directory, it kinda all has to
move together.

Note: this actually shows that dep files and `metadata_content` don't work
well together (we don't track it in dep files). That might actually be a
reasonable thing to do but we should probably document it somewhere.

Reviewed By: stepancheg

Differential Revision: D44712623

fbshipit-source-id: 3f86ae227f9c0086de05a5c6ab4a62bc0f9c4006