Prefer xcode-select over brew installing llvm
Summary:
macOS doesn't ship with homebrew by default, and we probably don't want to make brew a requirement for using or building buck2. This changes the recommendation to install clang to using Apple's provided Xcode command line tools, which can be installed with `xcode-select --install`.

Created from CodeHub with https://fburl.com/edit-in-codehub

Differential Revision: D44755988

fbshipit-source-id: c6e34e90bd68218fea85f2bc6fa03e7362ccc969