Add `prefer_remote` arg to `ctx.actions.run`
Summary:
**Context**

In some cases, we have dylibs at the bottom of the graph which are very big and which users want to link remotely to avoid downloading all object files. They still prefer to be able to link the top of the graph locally.

**Summary**

Add `prefer_remote` arg to `ctx.actions.run()`, so that the cxx layer can adjust behaviour to avoid local linking of specific targets under hybrid execution.

Reviewed By: krallin

Differential Revision: D44748418

fbshipit-source-id: 43788b891392b2ef1dd71644607a026bf44f49ed