Don't immediately unwrap protoc vendored paths
The intention of using environment variables is to override these paths.
With the code as written, this will unconditionally panic if
protoc_bin_vendored doesn't know where the protobuff compiler/includes
are, even if we're using an environment variable to set these paths.

With this change, we only unwrap when actually attempting to use the
path from `protoc_bin_vendored`.