Convert output-y buckevents to dap output events
Summary:
Without this conversion, at best these would get printed to stderr and
then vscode would ignore them and not surface them in any way. This way, the output will
end up in the debug console window. In normal execution of the debug-attach command, this
won't actually include anything, but it's helpful for getting error messages or for turning
on higher log levels.

Reviewed By: krallin

Differential Revision: D43941403

fbshipit-source-id: 2d361fdd7bf3fc8d36b08c11f127e02966f1f864