Add trait for subset of the debug adapter protocol and helpers
Summary:
This trait are direct encodings of the debug adapter protocol requests
and then some simple helpers for using it or encoding dap types. This almost doesn't
feel like a self-contained piece of code, but the next diff is
quite complex and just trying to get as much out of there as possible.

Reviewed By: krallin

Differential Revision: D43941372

fbshipit-source-id: ed044bbf220344ce10ef521ed9cafe838dd83c31