Add actual debugger server implementation
Summary:
Here's the main implementation of the buck2 debugger server,
connecting all the scaffolding we've put together to this point.

The documentation describes a lot of this, but here's a quick overview:

Events (from the dap client or triggered by starlark execution) all flow
through the single threaded ServerState loop. The main thing this does is
route DAP requests to the appropriate starlark DapAdapter, and then events
from starlark back to the DAP client. To do this it keeps track of the
ongoing starlark evaluations (and does other small things like assigning
them "pseudo thread ids").

In addition, the server sends out periodic debugger snapshots to all current
commands.

Reviewed By: krallin

Differential Revision: D43941373

fbshipit-source-id: 7197b878c268718a55a15afc8c19e8eb45047474