Don't immediately unwrap protoc vendored paths (#121)
Summary:
The intention of using environment variables is to override these paths. With the code as written, this will unconditionally panic if protoc_bin_vendored doesn't know where the protobuff compiler/includes are, even if we're using an environment variable to set these paths.

With this change, we only unwrap when actually attempting to use the path from `protoc_bin_vendored`.

Found as part of https://github.com/facebook/buck2/issues/120.

Pull Request resolved: https://github.com/facebook/buck2/pull/121

Reviewed By: cjhopman

Differential Revision: D44799077

Pulled By: ndmitchell

fbshipit-source-id: 56db51961166df0870cb4ad07a220abfc21c5008