Add show_includes dep file processor for MSVC
Summary:
This diff adds basic support to create dep files with MSVC.

[**Context**](https://learn.microsoft.com/en-us/cpp/build/reference/showincludes-list-include-files?view=msvc-170)
When the compiler comes to an include file during compilation, a message is output, as in this example:

```
Note: including file: d:\temp\1.h
Note: including file:  d:\temp\2.h
```
In this case, ```2.h``` was included from within ```1.h```, causing the indentation.

As in buck2 we don't care (yet) about which file includes which we strip out the prefix, normalize the paths and print to dep file.

Reviewed By: krallin

Differential Revision: D43244023

fbshipit-source-id: e9312672c55e64dd2b83545cf6f9b9acc61bf2cb