Support headers subtarget for Swift
Summary: To support `headers` subtarget, we need to expose Swift-related providers (SwiftDependencyInfo and SwiftPCMUncompiledInfo) to cxx layer. The main challenge here is that `propagated_preprocessor_args_cmd` is used to initialize SwiftPCMUncompiledInfo, but it's created inside of `cxx_library_parameterized`. So to handle this, we can pass a lamda as a parameter.

Reviewed By: chatura-atapattu

Differential Revision: D43951588

fbshipit-source-id: 5690ef3988a23da6fb4ae0f1c5f49994c19ced77