Make the Module always contain a DocItem
Summary:
The logic before was totally messed up - we only used it on Prelude, which was lucky, since in lots of cases it was probably wrong. If you have:

```
x = ""
```

That should mean you have a value `x` of type string. It actually mean we exported the entire string object with its methods. What's even worse, we had a test that checked we did that.

Reviewed By: stepancheg

Differential Revision: D44731844

fbshipit-source-id: 97ea0e78440f95451b5694c52df6eb44a9527a9b