buck2/prelude: don't append _test to go unittest package names
Summary:
Appending `_test` to go unittest packages names causes the source
locations embedded in the binary to also contain `_test`. This confuses
debuggers and code coverage processing tools as the paths don't actually exist.

In past folks just stripped away `_test` from paths, but that seems brittle
too.

It's unclear to me why we need the pakcage name to be <something>_test anyway.
Removing it from the build rule only makes the test that checked for the
explicit extra path fail. It does solve the code coverage issue for a bunch of
targets I tried.

Reviewed By: echistyakov

Differential Revision: D44896082

fbshipit-source-id: 6e94922ddc40325192e62214b075fb1e0c30cf20