Parameterize components by components
Summary:
OK, long story.

We should stop using `superconsole::State` because it is dynamically typed evil. Also because it does unnecessary restrictions like not be able to use non-static types in `State` (and we do not want to pull `AnyLifetime` here).

Do be able to do so efficiently (and for other reasons which are too long to explain here; TL;DR component now requires `Send`, and it is not easy to remove), we should allocate components on the stack instead of on the heap.

So for example, we want to invoke `Aligned` component as:

```
fn draw_unchecked(&self, ...) {
  let my_component = MyComponent { ... };
  Aligned::new(my_component).draw(...)
}
```

Differential Revision: D44814275

fbshipit-source-id: e60f8e071bf09e81f46c5f08a866bd5294ab7514