Extract class-to-src map + subtarget generation to a separate helper function
Summary:
This logic will need to be reused for Kotlin tests as well, so extracting to a separate helper function will make this a bit easier to use.

Ideally this would all go into `create_java_library_providers`, but things get a bit more complicated since there are subtargets being generated which need to be included in `DefaultInfo`. Could probably be done, but there's more refactoring that needs to be done first, so the approach in this diff feels reasonable for now.

Reviewed By: IanChilds

Differential Revision: D44927982

fbshipit-source-id: 00d9f4e40f780542716eced16007749c0490aa4a