buck2: split up daemon constraints
Summary:
I want to refactor this to allow us to add constraints that are only available
if the DaemonStateData was successfully created. This is what this does. I'm
moving trace io in it as a) a POC and b) because I'm going to move where
we create the daemon constraints and that'll become required then.

While in there, re-do how we express the TraceIoState. Right now, the daemon
can tell you its state is "existing", which makes no sense (it's either on or
off), instead it's only the client matcher that needs an enum (and we can
introduce that change now that the previous diffs in this stack are done).

Note: I think it might actually be a good idea to just return the error here
eventually.

Reviewed By: get9

Differential Revision: D44914243

fbshipit-source-id: 34556aab43dcb976d0c9dcbdb8ae4df300b357ce