buck2: decouple DaemonConstraints from their matcher
Summary:
We have the same struct to represent "what the daemon has" and "what we want
from the daemon", and that does not work great if we want to have more
fields, such as "I want any daemon that doesn't have daemon UUID XYZ in order
to force a restart".

So, separate those two things. The client side has its own matcher, and the
daemon side has its GRPC struct.

Reviewed By: get9

Differential Revision: D44879596

fbshipit-source-id: f37d33ee8a27de4d95ab4f482ba88a04adfb11ae