Do not use superconsole::State in TimedList
Summary:
`TimedList` component now grabs `&SuperConsoleState` in constructor instead of fetching it from the state.

So `TimedList` is statically typed now. No more errors because code compiles but does not work because forgot to put something in the state.

Code is a bit rough, `SuperConsoleState` is too much for `TimedList`. Can be cleaned up later, even as is I consider this version a big improvement.

Next step would be:
* use this pattern everywhere else in buck2
* remove `State` from `superconsole` and recommend this pattern as default
* similarly we can remove `DrawMode` from `superconsole`

Reviewed By: bobyangyf

Differential Revision: D44814677

fbshipit-source-id: c7a0ca14c71bff9c226cd13a49156cbb44cc7ae9