Drop superconsole::State
Summary:
Now all the state is passed explicitly to the components. State is no longer used.

State is inconvenient because it is not statically typechecked, so it is fragile, especially for components which requests different parts of state conditionally.

This diff stack addresses the issue https://github.com/facebookincubator/superconsole/issues/6

Next step would be removal of the state from `superconsole` crate, but before that, other users of `superconsole` need to be migrated.

Reviewed By: themarwhal

Differential Revision: D44815383

fbshipit-source-id: 5cd8caea3a9663ff8dfd171921167e4c24915d52