buck2: build_listener: error out on cycles (in the "default" backend)
Summary:
Like it says in the title. Not sure why I've seen an instance of a Buck2 being
stuck using up a bunch of CPU and memory in this code. Their build was super
small and yet this ran for a few minutes and ate 30 GB+ memory and I can't see
how that could be caused by anything but a cycle here.

This happens on a BRR, so this *might* be caused by eager deps (not sure at
all).

In any case, cycle here = OOM, so there is no reason not to error out instead.

Note that errors here are not blocking.

Reviewed By: stepancheg

Differential Revision: D44966727

fbshipit-source-id: ea9da2984bafc63f13d037371a8fe77057e735a7