chore: add boilerplate for hacking on buck2 with nix + direnv (#145)
Summary:
I've been using this to hack on and build `buck2` for a while now, since I often don't use `rustup` (and actually not using NixOS -- but Nix-On-Ubuntu-On-WSL2?!?!!) and figured it might be worth putting upstream at this point, with the official announcement being made. This will at least let any prospective users who have Nix get compiling real quick. It will also allow some future improvements to `flake.nix`, but...

Please note (especially lurking Nix users): this only allows you to use Nix to provision a development environment, one which is capable of running `cargo build`. There is not yet a "nix package" (AKA thing that installs inside `/nix/store`) for buck2 in this PR. I have one as part of `buck2-nix`, but it's out of date because some recent changes in mid/late February started causing a build failure. I'll open up a separate PR for that later and discuss it.

Please see the commit message for a longer version.

Pull Request resolved: https://github.com/facebook/buck2/pull/145

Reviewed By: iguridi

Differential Revision: D44908124

Pulled By: ndmitchell

fbshipit-source-id: 449c1f2bf2954a1def6a4eb8183e18c65839ca74