buck2: make sure killall doesn't hang on Windows
Summary:
Simpler version of D44831994.

Here we aren't displaying the parent list, just ensuring no infinite loops occur.

In the worst scenario, `killall` might not terminate a buck2 process because it incorrectly thinks it's an ancestor of the current process. However, several factors need to align for this to happen, making the likelihood very low.

Reviewed By: stepancheg

Differential Revision: D44999070

fbshipit-source-id: 4b6e013d49f3da8ae14476f7d8f55638a033351f