buck2: fix bulk delete hitting SQLite limits
Summary:
This does not currently work properly if you pass it more than 32K paths, at
which point it'll just error out

Unfortunately since we don't have a way to report materializer errors that gets
silently ignored! (that is however changing so we'll soon be able to fix this)

When used with `clean --stale`, the upshot is that we end up not deleting
things we are using from the DB, which predictably breaks everything :/

It is a bit embarrassing this took us this long to fix, but now that we log
soft errors to Scuba, it is rather obvious that this is what happened e.g.
in here:

https://fb.workplace.com/groups/buck2eng/posts/3137304493233696/?comment_id=3163719657258846

Reviewed By: lmvasquezg

Differential Revision: D45042250

fbshipit-source-id: b40a7a5a0c7c6622a53c54936272c644e6165e8d