buck2: bump sql schema version
Summary:
There is a big in clean-stale (fixed in the previous diff) that can leave
buck-out corrupted.

This diff is to ensure that if a buggy clean-stale executed, we don't reuse the
previous materializer state. To do that, we just bump the schema version, which
means that if you're using a Buck past this diff, you cannot be sharing your
materializer state with a Buck prior to this diff (and vice-versa).

This matters so that if a user leaves their repo checked out on an older rev
(and therefore an older Buck2), the overnight automated clean up does not
corrupt their state.

Reviewed By: lmvasquezg

Differential Revision: D45042414

fbshipit-source-id: 33eafbaa7ab245d5da6953c07adf8abc049c34d2