buck2: add a new type to represent a digest algorithm ... that we can use to hash
Summary:
What a weird title!

So, the reason we need this is because we want to use Keyed Blake3 internally.

But this means the algorithm is no longer just an enum of "Sha1, Sha256,
Blake3! Now, Keyed Blake3 needs to carry some data as well.

The problem is, we use DigestAlgorithmKind also for logging purposes sometimes
(e.g. when we store things in our materializer state database).

So, we need two types of digests algorithm here. The DigestAlgorithmKind is just
"ah yes this is keyed blake3 hash, but I don't have the key", which is what we
use in places where we no longer need to actually do hashing, and the
DigestAlgorihm is for when we actually require hashing.

The diff is a rather mechanical change, there's just one little API that needs
adding in `directory.rs` where we need to infer the DigestAlgorithm and not
just the DigestAlgorithmKind.

Reviewed By: bobyangyf

Differential Revision: D45005225

fbshipit-source-id: de4a31b4ddaf7804cf2616fa5d7ab7aa691717f9