Add the label to the project_external_debug_info method
Summary:
`project_external_debug_info` takes `ExternalDebugInfo`s, adds any additional artifacts, and projects them.

In the next diff, we need to add the `ctx.label` to `ExternalDebugInfo`, which requires us to pass the label to `project_external_debug_info`.

For simplicity of review, update all the call sites in this diff and pass the label.

The label is used in `ExternalDebugInfo` to later filter which debug infos we return under focused debugging.

Reviewed By: milend

Differential Revision: D44941274

fbshipit-source-id: 444960b72545ba8a21e51ba19c1886ffcad23939