Add the label to ExternalDebugInfo tset
Summary:
Now that we pass the label down from all callsites, create an ExternalDebugInfo record that is used in the ExternalDebugInfoTSet, and all the label to it.

The projection will still project all debug artifacts and will maintain the performance of using a projection for the debug info as in D41920011.

Up the stack, when performing focused debugging, we will have to perform a traverse and filter which debug infos to include.

Reviewed By: milend

Differential Revision: D44941275

fbshipit-source-id: 6232518fa93b09a31060638e5104453ae603cad2