buck2: restarter: actually apply it
Summary:
The restarter keeps track of why we want to restart. This actually makes it do something
by plugging it into our daemon connection constraints.

This is how we can get behavior such as "get me a new daemon and restart the command".

We implement this via constraints that say " not this daemon" or "not this materializer state"
instead of just "kill the daemon then restart" so that if you have e.g. 2 parallel commands
that both hit an issue and decide to restart, only one will kill the daemon and the second one
will be happy to just reuse the restarted daemon.

Reviewed By: get9

Differential Revision: D44997255

fbshipit-source-id: 0738c3e102c92ba67567f33fdef216d53a7ed26e