buck2: ExitResult: let it carry stdout
Summary:
Like it says in the title. This allows us to do restarts better, because we can
prevent "final" output from being produced (which would prevent us from
restarting).

This is particularly useful in `build` where we might print the build report
that way. It is not useful in e.g. queries where we stream to stdout, but
those are rather less likely to start hitting a condition that would warrant
a daemon restart after they start producing output, so that's probably OK.

Reviewed By: get9

Differential Revision: D45002630

fbshipit-source-id: ed4ecc83f8c2c790fc69d3238983fa826b0b18f0