buck2: introduce restarter
Summary:
Like it says in the title. We pass this down to the CLI to collect some state (thats coming into the next diffs)
and then we'll use that to decide whether we want to restart the command.

Since we don't want to reuse the same trace id across restarts, we also handle that here.

Reviewed By: get9

Differential Revision: D44997252

fbshipit-source-id: 840731da949075173fd5e3b3fffd7376c030995a