buck2: implement restarter
Summary:
Like it says in the title. Actually give the restarter the ability to observe what we did
so that it can decide whether we should restart the command.

This pattern-matches off of our existing implementation that we use for exit codes
since the reasons we set an exit code and the reasons we decide to restart are
usually fairly similar.

Reviewed By: get9

Differential Revision: D44997254

fbshipit-source-id: 60bad937aac910c52119a99327eca8611db879dd