buck2: build: pass stdout via the ExitResult
Summary:
There is a little bit of output that build wants to produce over stdout, in
particular the build report (if you request it).

However, when doing auto-restart, we require stdout to be empty (since
otherwise we might be producing invalid output), so that means we can't restart
those commands successfully.

To fix that, we now pass stdout via the ExitResult (which is fairly similar to
what we do with passing a `Exec` request via the `ExitResult`), which allows us
to avoid printing it when we want to restart.

Reviewed By: get9

Differential Revision: D45002625

fbshipit-source-id: 4cd638d4e2443e1d312849b7bcf9c372314e351b