buck2: materializer: make materializer_invalidate_error a hard error
Summary:
Like it says in the title. As per our logs this is only ever caused by the
crash fixed earlier in this stack, so let's make it a blocking error.

I do expect that this could also trigger if the user has run out of disk
space, but then in that case it's still more desirable to fail.

Reviewed By: christolliday

Differential Revision: D45044736

fbshipit-source-id: 17afc553b03aefb17c3d13e7a9a805d1bbca01b1