buck2: install: use `MaterializationContext`
Summary:
This does a bit of dedupe that we know improves perf in `build` in some
scenarios, so might as well use it here.

Another nice effect of it is that it avoids notifying our critical path
calculation of the same output multiple times, which itself isn't fundamentally
terrible (though we currently soft error on it), but actually could mean you
get the wrong timings (you want the timing reported by the first
materialization not the last one).

Reviewed By: bobyangyf

Differential Revision: D45091029

fbshipit-source-id: 5f4859dad349a67c1bedc2dbfd7dc55a73d92f75