For loop in bytecode
Summary:
Change how `for` loop works in bytecode.

Before this diff there was `ForLoop` instruction, which called evaluation recursively.

`Continue` and `Break` instructions just returned from that recursion back to `ForLoop` instruction.

The downside is that each nested for loop increased rust stack size.

In new implementation, iterator value is stored as regular value in on the stack. Four loop-related instructions update the iterator value and jump to the appropriate instruction address. No recursion.

This machinery is also prerequisite to proper GC in starlark: now that iterator variable is stored on the stack, we can trace it during GC. So we can do GC while iteration is active, which is not possible before this diff.

Reviewed By: ndmitchell

Differential Revision: D45034981

fbshipit-source-id: 6fdafbf7b3f76af8cf4fa8a78059132b0cf3714c