Remove unorganized comments
Summary: It's true that these "`// used in xxx`" comments are correct, but they are not consistent among the other lines. They should be consistent in either way, putting comments on every line or no comment. By this diff, we are removing them as the comment doesn't have any important meaning (you can know it by reading the code.)

Reviewed By: stepancheg

Differential Revision: D45136813

fbshipit-source-id: 130f1956f1ec014fd2b3bef50edb0cf2724553da